### 2019-05-23 #305 @krlmlr
Submission: tic
Submitting Author: Kirill Müller (@krlmlr)
Repository: [ropenscilabs/tic](https://github.com/ropenscilabs/tic)
Version submitted: 0.2.13.0916
Version accepted: 0.3.0.9004

---

## `DESCRIPTION` file

```
Package: tic
Title: Tasks Integrating Continuously: CI-Agnostic Workflow Definitions
Version: 0.2.13.9016
    person("Kirill", "M\u00fcller", email = "krlmlr+r@mailbox.org", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1416-3412")),
    person("Patrick", "Schratz", email = "patrick.schratz@gmail.com", role = "aut", comment = c(ORCID = '0000-0003-0748-6624')),
    person("Mika", "Braginsky", email = "mika.br@gmail.com", role = c("aut")),
    person("Karthik", "Ram", email = "karthik.ram@gmail.com", role = c("aut")),
    person("Jeroen", "Ooms", email = "jeroen.ooms@stat.ucla.edu", role = c("aut")),
    person("rOpenSci", role = c("fnd"))
  )
    Provides a way to describe common build and deployment workflows for R-based
    projects: packages, websites (e.g. blogdown, pkgdown), or data processing
    (e.g. research compendia).  The recipe is described independent of the
    continuous integration tool used for processing the workflow (e.g.
    'Travis CI' or 'AppVeyor').
    backports,
    cli,
    crayon,
    memoise,
    remotes,
    R6,
    withr
    base64enc,
    bookdown,
    callr,
    covr,
    devtools,
    drat,
    fansi,
    git2r,
    openssl,
    rcmdcheck,
    travis (>= 0.2.11.9001),
    usethis
    r-lib/rcmdcheck,
    ropenscilabs/travis
ByteCompile: No
Collate:
    'ci.R'
    'appveyor.R'
    'base64.R'
    'dsl-storage.R'
    'dsl.R'
    'git2r.R'
    'keys.R'
    'local.R'
    'macro.R'
    'macro-package-checks.R'
    'macro-pkgdown.R'
    'macro-bookdown.R'
    'mock.R'
    'print.R'
    'repo.R'
    'run.R'
    'stage.R'
    'steps-base.R'
    'steps-bookdown.R'
    'steps-code.R'
    'steps-drat.R'
    'steps-git.R'
    'steps-install.R'
    'steps-rcmdcheck.R'
    'steps-pkgdown.R'
    'steps-ssh.R'
    'steps-write-text-file.R'
    'tic-package.R'
    'travis.R'
    'use_tic.R'
    'utils.R'
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under:

> Provides a better interface to CI systems: define CI/CD using R code instead of YAML files. Work supported by rOpenSci.

> R package developers, R + Git users who would like to run R code automatically on code update and deploy artifacts.

> `language: r` support in Travis CI, the [r-appveyor](https://github.com/krlmlr/r-appveyor/) project for AppVeyor. This project aims at integrating these and other services using a common DSL.

---

CC @pat-s @maelle @karthik
