### 2019-05-29 #305 @annakrystalli
### Editor checks:

---

#### Editor comments

Hello again @krlmlr,

I've completed the editors checks, see below for issues raised. Feel free to address/respond to any issues raised.

Can I also please ask that you add the rOpenSci review badge to the README?

In the meantime, I'll start looking for reviewers!

***

##### checks

The checks throw up an example which isn't working:

``` r
#> ── R CMD check results ─────────────────────────────── tic 0.2.13.9016 ────
#> Duration: 54.8s
#>
#> ❯ checking examples ... ERROR
#>   Running examples in ‘tic-Ex.R’ failed
#>   The error most likely occurred in:
#>
#>   > base::assign(".ptime", proc.time(), pos = "CheckExEnv")
#>   > ### Name: step_do_push_deploy
#>   > ### Title: Step: Perform push deploy
#>   > ### Aliases: step_do_push_deploy
#>   >
#>   > ### ** Examples
#>   >
#>   > dsl_init()
#>   [32m✔[39m Creating a blank tic stage configuration
#>   >
#>   > get_stage("deploy") %>%
#>   +   add_step(step_do_push_deploy(path = "docs"))
#>   Error: Error evaluating the step argument of add_step(), expected an object of class TicStep.
#>   Original error: The working directory is not in a git repository
#>   Execution halted
#>
#> 1 error ✖ | 0 warnings ✔ | 0 notes ✔
#> Error: R CMD check found ERRORs
```

##### tests

`devtools:test()` is throwing up some warnings for me (well the same warning in a couple of locations):

``` r
pkg_dir <- "/Users/Anna/Documents/workflows/rOpenSci/editorials/tic"
devtools::test(pkg_dir)
⠙ |   1   1   | test-integration-git.R
✔ |   1   1   | test-integration-git.R [0.8 s]
#> ───────────────────────────────────────────────────────────────────────────────────
#> test-integration-git.R:19: warning: integration test: git
#> `recursive` is deprecated, please use `recurse` instead
#> ───────────────────────────────────────────────────────────────────────────────────
#>
⠏ |   0       | test-integration-package.R┌───────────────────────────────────────┐
#> │                                       │
#> │   integration test: package failure   │
#> │                                       │
#> └───────────────────────────────────────┘
#>
#> Error : A step failed in stage "script": script.
#>
⠹ |   1   2   | test-integration-package.R
✔ |   1   2   | test-integration-package.R [6.2 s]
#> ───────────────────────────────────────────────────────────────────────────────────
#> test-integration-package-failure.R:9: warning: integration test: package failure
#> `recursive` is deprecated, please use `recurse` instead
#>
#> test-integration-package-failure.R:9: warning: integration test: package failure
#> `recursive` is deprecated, please use `recurse` instead
#> ───────────────────────────────────────────────────────────────────────────────────
#>
⠏ |   0       | test-integration-package.R┌───────────────────────────────┐
#> │                               │
#> │   integration test: package   │
#> │                               │
#> └───────────────────────────────┘
⠹ |   1   2   | test-integration-package.R
✔ |   1   2   | test-integration-package.R [5.6 s]
#> ───────────────────────────────────────────────────────────────────────────────────
#> test-integration-package.R:9: warning: integration test: package
#> `recursive` is deprecated, please use `recurse` instead
#>
#> test-integration-package.R:9: warning: integration test: package
#> `recursive` is deprecated, please use `recurse` instead
#> ───────────────────────────────────────────────────────────────────────────────────
#>
⠏ |   0       | test-repo
⠋ |   1       | test-repo
⠙ |   2       | test-repo
⠹ |   3       | test-repo
⠸ |   4       | test-repo
✔ |   4       | test-repo [3.9 s]
#>
⠏ |   0       | test-stages.R
✔ |   2       | test-stages.R
#>
⠏ |   0       | tasks-base
✔ |   6       | tasks-base
#>
⠏ |   0       | test-utils.R
✔ |   4       | test-utils.R
#>
#> ══ Results ════════════════════════════════════════════════════════════════════════
#> Duration: 24.3 s
#>
#> OK:       64
#> Failed:   0
#> Warnings: 5
#> Skipped:  0
```

##### goodpractice

Output from `goodpractice::gp(pkg_dir)`

```
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 51% of code lines are covered by test
    cases.

    R/base64.R:15:NA
    R/base64.R:16:NA
    R/base64.R:17:NA
    R/base64.R:18:NA
    R/base64.R:31:NA
    ... and 368 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80
    characters wide. Try make your lines shorter than 80 characters

    R/appveyor.R:24:1
    R/ci.R:45:1
    R/dsl-storage.R:81:1
    R/dsl-storage.R:85:1
    R/dsl-storage.R:110:1
    ... and 88 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import
    only the specific functions you need.
───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

##### spelling

A few candidates for genuine typos identified through `devtools::spell_check()`

```
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD              FOUND IN
beim              tic.Rmd:221
builded           NEWS.md:195
commited          deployment.Rmd:166
constitutent      dsl.Rd:37
interfer          tic.Rmd:129
undefinied        NEWS.md:227
ususally          build-lifecycle.Rmd:20
```

Finally, picking up from the coverage issue picked up by `gp()`, `coverage()` breaks down the problem areas in a bit more detail:

```
tic Coverage: 51.93%
R/base64.R: 0.00%
R/keys.R: 0.00%
R/macro-bookdown.R: 0.00%
R/macro-pkgdown.R: 0.00%
R/mock.R: 0.00%
R/print.R: 0.00%
R/steps-bookdown.R: 0.00%
R/steps-drat.R: 0.00%
R/steps-pkgdown.R: 0.00%
R/steps-ssh.R: 0.00%
R/steps-write-text-file.R: 0.00%
R/use_tic.R: 0.00%
R/ci.R: 36.84%
R/steps-install.R: 42.31%
R/local.R: 46.15%
R/steps-code.R: 46.15%
R/utils.R: 48.00%
R/macro-package-checks.R: 60.71%
R/git2r.R: 66.67%
R/dsl-storage.R: 72.41%
R/steps-rcmdcheck.R: 73.68%
R/stage.R: 85.92%
R/dsl.R: 87.18%
R/run.R: 88.46%
R/steps-git.R: 90.79%
R/repo.R: 90.91%
R/steps-base.R: 100.00%
```

In general, we require atleast 75% test coverage. Is there a reason coverage is at ~50% and that a number of `R/` scripts aren't covered at all?

### Comments for reviewers

- There is a NOTE that is output from an `R CMD check` which is run as part of `test-integration-package.R` on a fake package.

    ```r
    #> [36m── R CMD check results ─────────────────── ticpkg457a3a5e2a7apkg 0.0.0.9000 ────[39m
    #> Duration: 3s
    #>
    #> [34m❯ checking top-level files ... NOTE[39m
    #>   Non-standard file/directory found at top level:
    #>     ‘tic.R’
    #>
    ```
  You can safely ignore it. It's just the result of the tic file in the fake package not having been added to the `.Rbuildignore` file. I'm just flagging it because I got confused as there's a lot of output from the tests. Thanks @maelle for help trouble shooting!

- Also can I draw your attention to the actual test at the end of `test-integration-package.R` which is a simple `expect_true(TRUE)` and invite you to consider whether a more specific test could be more appropriate?

---

Reviewers:
Due date:
