### 2019-06-03 165f770
- Condition example on presence of Git repository.
