+++ man/step_do_push_deploy.Rd
@@ -43,8 +43,10 @@ the deployment is safely applied to the current tip.
 \examples{
 dsl_init()
 
-get_stage("deploy") \%>\%
-  add_step(step_do_push_deploy(path = "docs"))
+if (rlang::is_installed("git2r") && git2r::in_repository()) {
+  get_stage("deploy") \%>\%
+    add_step(step_do_push_deploy(path = "docs"))
+}
 
 dsl_get()
 }
