+++ vignettes/deployment.Rmd
@@ -38,7 +38,7 @@ knitr::opts_chunk$set(
 
 [`pkgdown`](https://github.com/r-lib/pkgdown) is a R package which builds a documentation wrapper-site of a R package.
 It collects all the vignettes, function references and metadata information of a package and presents it in an eye-appealing web version.
-Without a `pkgdown` site, these package resources need either be collected from various places (Github repository, CRAN function reference) or do even require a local installation of the package if vignettes are placed in custom locations. 
+Without a `pkgdown` site, these package resources need either be collected from various places (GitHub repository, CRAN function reference) or do even require a local installation of the package if vignettes are placed in custom locations. 
 Therefore, it has become a quasi-standard to provide a _pkgdown_ site for a R package.
 However, it is tedious to update the site on every commit and check whether something has changed.
 On the other hand, the presence of a _pkgdown_ site is only really valuable to the user if the content reflects the most recent state of the package.
