### 2019-06-03 02f3485
- Add `tic.R` to `.Rbuildignore` for internal tests.
