### 2019-06-03 b3cd2f3
Style
- Strip long source code lines.
- Add review badge.
