+++ tests/testthat/test-integration-git-race-path-branch-orphan.R
@@ -1,6 +1,6 @@
 context("test-integration-git-race-path-branch-orphan.R")
 
-test_that("integration test: git race condition with path and branch arguments set, deployment with orphaning", {
+test_that("integration test: git race condition + path + branch + orphaning", {
   cli::cat_boxx("integration test: git race path branch orphan")
 
 
@@ -28,17 +28,16 @@ test_that("integration test: git race condition with path and branch arguments s
   package_path <- tmp("package")
   git2r::clone(bare_repo_path, package_path)
 
+  tic_r <- c(
+    'get_stage("deploy") %>%',
+    '  add_code_step(writeLines(sort(dir(pattern = "^clone[.]txt$", recursive = TRUE)), "deploy/dir.txt")) %>%',
+    paste0('  add_step(step_push_deploy(path = "deploy", branch = "deploy-branch", remote_url = "', bare_repo_path, '"))')
+  )
+
   cat("\n")
   withr::with_dir(
     package_path, {
-      writeLines(
-        c(
-          'get_stage("deploy") %>%',
-          '  add_code_step(writeLines(sort(dir(pattern = "^clone[.]txt$", recursive = TRUE)), "deploy/dir.txt")) %>%',
-          paste0('  add_step(step_push_deploy(path = "deploy", branch = "deploy-branch", remote_url = "', bare_repo_path, '"))')
-        ),
-        "tic.R"
-      )
+      writeLines(tic_r, "tic.R")
       writeLines("^tic\\.R$", ".Rbuildignore")
       git2r::config(user.name = "tic", user.email = "tic@pkg.test")
       dir.create("deploy")
@@ -68,7 +67,9 @@ test_that("integration test: git race condition with path and branch arguments s
   withr::with_dir(
     package_path_3, {
       writeLines("clone-contents", "clone.txt")
-      git2r::config(user.name = "tic-clone-2", user.email = "tic-clone-2@pkg.test")
+      git2r::config(
+        user.name = "tic-clone-2", user.email = "tic-clone-2@pkg.test"
+      )
       git2r::add(path = ".")
       git2r::commit(message = "Edit clone.txt")
       git2r::push()
@@ -103,7 +104,9 @@ test_that("integration test: git race condition with path and branch arguments s
       system2("git", "reset origin/deploy-branch --hard")
       expect_equal(length(git2r::commits()), 1)
       expect_false(file.exists("clone.txt"))
-      expect_equal(readLines("dir.txt"), sort(dir(package_path, pattern = "^clone[.]txt$", recursive = TRUE)))
+      expect_equal(
+        readLines("dir.txt"), sort(dir(package_path, pattern = "^clone[.]txt$", recursive = TRUE))
+      )
     }
   )
 
@@ -125,7 +128,9 @@ test_that("integration test: git race condition with path and branch arguments s
       system2("git", "reset origin/deploy-branch --hard")
       expect_equal(length(git2r::commits()), 1)
       expect_false(file.exists("clone.txt"))
-      expect_equal(readLines("dir.txt"), sort(dir(package_path_2, pattern = "^clone[.]txt$")))
+      expect_equal(
+        readLines("dir.txt"), sort(dir(package_path_2, pattern = "^clone[.]txt$"))
+      )
     }
   )
 
@@ -148,7 +153,9 @@ test_that("integration test: git race condition with path and branch arguments s
       expect_equal(length(git2r::commits()), 1)
       expect_false(file.exists("clone.txt"))
       print(readLines("dir.txt"))
-      expect_equal(readLines("dir.txt"), sort(dir(package_path_3, pattern = "^clone[.]txt$", recursive = TRUE)))
+      expect_equal(
[…]
