+++ R/print.R
@@ -2,7 +2,10 @@
 #' @export
 print.TicStages <- function(x, ...) {
   if (all(vlapply(x, stage_is_empty))) {
-    cat_bullet("No steps defined in any stage", bullet = "info", bullet_col = "green")
+    cat_bullet(
+      "No steps defined in any stage",
+      bullet = "info", bullet_col = "green"
+    )
   } else {
     lapply(x, print, omit_if_empty = TRUE)
   }
