+++ R/macro-package-checks.R
@@ -60,7 +60,8 @@ do_package_checks <- function(...,
     )
 
   if (isTRUE(codecov)) {
-    #' 1. A call to [covr::codecov()] in the `"after_success"` stage (only if the `codecov` flag is set)
+    #' 1. A call to [covr::codecov()] in the `"after_success"` stage
+    #'    (only if the `codecov` flag is set)
     get_stage("after_success") %>%
       add_code_step(covr::codecov(quiet = FALSE))
   }
