+++ tests/testthat/test-integration-git-race-path-branch-orphan.R
@@ -105,7 +105,8 @@ test_that("integration test: git race condition + path + branch + orphaning", {
       expect_equal(length(git2r::commits()), 1)
       expect_false(file.exists("clone.txt"))
       expect_equal(
-        readLines("dir.txt"), sort(dir(package_path, pattern = "^clone[.]txt$", recursive = TRUE))
+        readLines("dir.txt"),
+        sort(dir(package_path, pattern = "^clone[.]txt$", recursive = TRUE))
       )
     }
   )
@@ -129,7 +130,8 @@ test_that("integration test: git race condition + path + branch + orphaning", {
       expect_equal(length(git2r::commits()), 1)
       expect_false(file.exists("clone.txt"))
       expect_equal(
-        readLines("dir.txt"), sort(dir(package_path_2, pattern = "^clone[.]txt$"))
+        readLines("dir.txt"),
+        sort(dir(package_path_2, pattern = "^clone[.]txt$"))
       )
     }
   )
@@ -154,7 +156,8 @@ test_that("integration test: git race condition + path + branch + orphaning", {
       expect_false(file.exists("clone.txt"))
       print(readLines("dir.txt"))
       expect_equal(
-        readLines("dir.txt"), sort(dir(package_path_3, pattern = "^clone[.]txt$", recursive = TRUE))
+        readLines("dir.txt"),
+        sort(dir(package_path_3, pattern = "^clone[.]txt$", recursive = TRUE))
       )
     }
   )
