### 2019-06-03 366a4d2
Shorten lines
