### 2019-06-03 34e5d06
- Exclude code that can only run interactively or in a CI from coverage.
