### 2019-06-03 d4146df
- Test utils and printing.
