+++ tests/testthat/test-base64.R
@@ -1,0 +1,4 @@
+test_that("roundtrip serialization", {
+  x <- runif(10)
+  expect_identical(base64unserialize(base64serialize(x)), x)
+})
