+++ R/utils.R
@@ -37,27 +37,6 @@ get_deps_from_code <- function(call) {
   }
 }
 
-verify_install <- function(...) {
-  pkg_names <- c(...)
-  lapply(pkg_names, verify_install_one)
-}
-
-verify_install_one <- function(pkg_name) {
-  if (!package_installed(pkg_name)) {
-    utils::install.packages(pkg_name)
-    if (!package_installed(pkg_name)) {
-      stopc(
-        "Error installing package ", pkg_name, " or one of its dependencies."
-      )
-    }
-  }
-}
-
-package_installed <- function(pkg_name) {
-  path <- system.file("DESCRIPTION", package = pkg_name)
-  file.exists(path)
-}
-
 format_traceback <- function(top = NULL, bottom = parent.frame()) {
   paste(format(trace_back(top, bottom)), collapse = "\n")
 }
