### 2019-06-03 #305 @krlmlr
Thanks. I've addressed the editorial comments in v0.2.13.9017:

- Condition example on presence of Git repository.
- Fix typos.
- Update wordlist.
- Add `tic.R` to `.Rbuildignore` for internal tests to remove spurious warnings in test output.
- Add review badge.
- Reformat long source code lines.
- I don't think I can do much better than `expect_true(TRUE)`, added a comment.
- Exclude code that can only run interactively or in a CI from coverage.
- Test utils and printing.

The comments regarding coverage helped identify a smell in the API, thanks! Working on improving it, will report here.

{rlang} and {cli} are sufficiently safe to be imported as a whole, {backports} really needs to be imported as a whole.
