### 2019-06-03 81094e2
Move use_travis_deploy() from travis, unexport get_public_key() and encode_private_key()
