+++ R/setup.R
@@ -1,0 +1,97 @@
+# This code can only run interactively
+# nocov start
+
+#' Set up deployment for Travis CI
+#'
+#' Creates a public-private key pair,
+#' adds the public key to the GitHub repository via [travis::github_add_key()],
+#' and stores the private key as an encrypted environment variable in Travis CI
+#' via [travis::travis_set_var()], possibly in a different repository.
+#' See [step_setup_ssh()] or [step_install_ssh_keys()]
+#' for installing such a key during a Travis CI build.
+#'
+#' @param travis_repo `[string]`\cr
+#'   The Travis CI repository to add the private key to, default:
+#'   the GitHub repo to which the public deploy key is added.
+#' @param name `[string]`\cr
+#'   The private environment variable on Travis CI to write the private key to,
+#'   default: `"id_rsa"`.  Use this value as `name` argument to
+#'   [step_setup_ssh()] or [step_install_ssh_keys()].
+#'
+#' @export
+use_travis_deploy <- function(travis_repo = NULL, name = "id_rsa") {
+
+  # authenticate on github and travis and set up keys/vars
+  if (!is_installed("openssl")) {
+    stopc("Please install the openssl package.")
+  }
+
+  # generate deploy key pair
+  key <- openssl::rsa_keygen()  # TOOD: num bits?
+
+  # encrypt private key using tempkey and iv
+  pub_key <- get_public_key(key)
+  private_key <- encode_private_key(key)
+
+  info <- travis::github_info()
+  repo <- travis::github_repo(info)
+  if (is.null(travis_repo)) {
+    travis_repo <- repo
+  }
+
+  # add to GitHub first, because this can fail due to missing org permissions
+  title <- paste0(
+    "travis+tic", if (repo == travis_repo) "" else (paste0(" for ", repo))
+  )
+  travis::github_add_key(pub_key, title = title, info = info)
+
+  travis::travis_set_var(name, private_key, public = FALSE, repo = travis_repo)
+
+  cli::cat_bullet(
+    bullet = "tick", bullet_col = "green",
+    paste0(
+      "Successfully added private deploy key to ",
+      travis_repo, " as secure environment variable ", name, " to Travis CI."
+    )
+  )
+}
+
+#' Get public RSA key
+#'
+#' Extracts the public key from a public-private key pair
+#' generated by [openssl::rsa_keygen()].
+#' This key can be installed as a deploy key for a GitHub repository.
+#'
+#' @param key RSA key, as returned by `openssl::rsa_keygen()`
+#' @keywords internal
+#' @noRd
+get_public_key <- function(key) {
+  as.list(key)$pubkey
+}
+
+#' Encode a private RSA key
+#'
+#' Extracts the private key from a public-private key pair
+#' generated by [openssl::rsa_keygen()], and encodes it in
+#' base64 encoding.
+#' This key can be stored in the `id_rsa` private environment variable on Travis CI,
+#' from where [step_install_ssh_keys()] will pick it up to provide the CI process
+#' with deployment rights.
+#'
+#' @inheritParams get_public_key
+#' @seealso [step_install_ssh_keys()] [step_test_ssh()] [step_setup_ssh()]
+#' @keywords internal
+#' @noRd
+encode_private_key <- function(key) {
+  conn <- textConnection(NULL, "w")
+  openssl::write_pem(key, conn, password = NULL)
+  private_key <- textConnectionValue(conn)
+  close(conn)
+
+  private_key <- paste(private_key, collapse = "\n")
+
+  openssl::base64_encode(charToRaw(private_key))
+}
+
+# This code can only run interactively
+# nocov end
