+++ R/use_tic.R
@@ -43,6 +43,17 @@ use_tic <- function(quiet = FALSE) {
     paste0("Using active project ", usethis::ui_value(path))
   )
 
+  repo_type <- detect_repo_type()
+
+  if (needs_deploy(repo_type) && !is_installed("openssl")) {
+    cli::cat_rule(col = "red")
+    stopc(
+      "`use_tic()` needs the `openssl` package to set up deployment, ",
+      'please install using install.packages("openssl").'
+    )
+  }
+
+
   #' @details
   #' The preparation consists of the following steps:
   #' 1. If necessary, create a GitHub repository via [usethis::use_github()]
@@ -86,7 +97,6 @@ use_tic <- function(quiet = FALSE) {
   #' 1. Create a default `appveyor.yml` file
   #'    (depending on repo type, overwrite after confirmation
   #'    in interactive mode only)
-  repo_type <- detect_repo_type()
   if (needs_appveyor(repo_type)) use_appveyor_yml()
 
   #' 1. Create a default `tic.R` file depending on the repo type
@@ -102,7 +112,7 @@ use_tic <- function(quiet = FALSE) {
     ),
     col = "green"
   )
-  if (needs_deploy(repo_type)) travis::use_travis_deploy()
+  if (needs_deploy(repo_type)) use_travis_deploy()
 
   cli::cat_boxx(
     c(
