+++ NAMESPACE
@@ -39,8 +39,6 @@ export(do_pkgdown)
 export(dsl_get)
 export(dsl_init)
 export(dsl_load)
-export(encode_private_key)
-export(get_public_key)
 export(get_stage)
 export(install)
 export(prepare_all_stages)
@@ -69,6 +67,7 @@ export(step_setup_ssh)
 export(step_test_ssh)
 export(step_write_text_file)
 export(use_tic)
+export(use_travis_deploy)
 import(backports)
 import(cli)
 import(rlang)
