+++ DESCRIPTION
@@ -65,7 +65,6 @@ Collate:
     'dsl.R'
     'git2r.R'
     'install.R'
-    'keys.R'
     'local.R'
     'macro.R'
     'macro-package-checks.R'
@@ -75,6 +74,7 @@ Collate:
     'print.R'
     'repo.R'
     'run.R'
+    'setup.R'
     'stage.R'
     'steps-base.R'
     'steps-bookdown.R'
