### 2019-06-03 af24f26
Update documentation to reflect tic::use_travis_deploy()
