+++ man/step_push_deploy.Rd
@@ -33,6 +33,9 @@ commits, and pushes.
 Combines \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}} with \code{checkout = FALSE} and
 a suitable \code{orphan} argument,
 and \code{\link[=step_do_push_deploy]{step_do_push_deploy()}}.
+
+Deployment usually requires setting up SSH keys with
+\code{\link[=use_tic]{use_tic()}} or \code{\link[=use_travis_deploy]{use_travis_deploy()}}.
 }
 \details{
 Setup and deployment are combined in one step,
