### 2019-06-03 #305 @krlmlr
That API smell is removed in v0.2.13.9018. Looking forward to your reviews!
