### 2019-06-03 6af3625
- Building pkgdown site succeeds if `docs/` directory is missing (#173, r-lib/pkgdown#1050).
Closes #173.
