+++ R/steps-pkgdown.R
@@ -11,7 +11,9 @@ BuildPkgdown <- R6Class(
 
     run = function() {
       remotes::install_local(".")
-      pkgdown::clean_site()
+      if (dir.exists("docs")) {
+        pkgdown::clean_site()
+      }
       do.call(
         pkgdown::build_site, c(list(preview = FALSE), private$pkgdown_args)
       )
