### 2019-06-13 6ca690d
- Fix `R CMD build` and `R CMD check` switches on AppVeyor.
