### 2019-06-15 4d152a0
Ignore vignettes dir
