### 2019-06-18 ca3e6f8
add a note why pkgdown is only build on Travis
