### 2019-06-19 #305 @annakrystalli
Hey Laura,

You can find the template here: https://devguide.ropensci.org/reviewtemplate.html

Sent from my iPhone

> On 19 Jun 2019, at 18:08, Laura DeCicco <notifications@github.com> wrote:
>
> Can someone remind me where to find the blank reviewer templates?
>
