### 2019-06-23 #305 @maxheld83
As discussed with editor @annakrystalli, I am the maintainer of the [ghactions](https://r-lib.github.io/ghactions/) package for R workflow automation with GitHub Actions.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

CI/CD is extremely important for open and reproducible research, and the *tic* package is an incredibly ambitious, generous and comprehensive initiative to make it easier for more people to adhere to these best practices.

The promotion of CI/CD for *non-package* projects is particular helpful for research purposes, and somewhat unaddressed so far.

The package is thoroughly documented, with great introductory vignettes for newcomers.
I checked only a relatively small subset of the feature set, but found the UI/UX pretty smooth and didn't run into any problems.

So: definetely accept this great contribution!

I *do* have some reservations towards the *agnostic philosophy* of tic, though this isn't something I would bet on, let alone to reject this.
Users should decide for themselves what suits their needs, and time will tell which approaches work, and which won't.

Perhaps my thoughts can still be valuable for the authors.

I'm a bit skeptical as to whether it's a good idea to wrap *yet* another layer on top of the (usually) `*.yml`s used to express workflows on different CI services.
These `*.yml`s, after all, are *already* abstractions over (bash?) scripts run by the services, and *are* meant as the UI for users.

There is an obvious appeal to abstracting over different providers, and for making this so smooth, but there are also some potential downsides:

- The complexity of managing CI/CD workflows grows considerably; now there's *another* abstraction layer which may have bugs or unexpected behavior.
- This will, among other things, require the authors to keep their package up to date with any upstream changes to the `*.yml`s by the CI/CD service providers.
  (A similar problem appears to be raised in [#171](https://github.com/ropenscilabs/tic/issues/171))
- It's not obvious whether the abstraction (the "steps") actually *do* or *will* map elegantly to all providers now or in the future.
  For a counter-example, GitHub Actions *doesn't* know "steps."
  (GitHub Actions *can* be used for CI/CD, but it's still pretty new and weird, so this isn't a very relevant limitation right now -- and maybe it could be mapped).
- It's not obvious that the interface of tic is more obvious or user-friendly than writing a `*.yml`, which isn't really *that* hard.
  On the one hand, for easy CI/CD workflows, one can usually just copy/paste a `*.yml` (plus some GitHub PAT action, admittedly).
  For very custom/complicated CI/CD workflows, on the other hand, learning the R6 class system of tic may be even more complicated than adding the logic of some providers `*.yml`.
- *If* wrapping `*.yml`s in another abstraction / DSL, one might wonder whether R is, in fact, the right language and level at which to do this.
  Arguably, writing service-agnostic workflows could be such a big undertaking (which the tic team seems to have largely covered, incredibly!), that it might be better organised under a "bigger tent" with a larger (non-R) community, that then simply gets wrapped for R convenience.
  Absent such a bigger community, I hope that the authors will be able to cope with the maintenance load imposed by this (rapidly?) evolving/consolidating marketplace of tools.
- Alternatively, it would be interesting to see whether other communities (Python, Julia?) have *also* developed their own CI/CD abstraction/DSL.
- Lastly, one might wonder whether it's actually a good idea *pedagogically*, so to speak, to isolate R users from anything non-R (bash scripting, `*.yml`).
  I understand that many R users are *not* software developers and have limited patience and time for yet more tooling, but by wrapping ever more tools in R, we risk keeping (new) users relatively unprepared for the times when they *will* need to use outside (standard) tooling.

I've learned CI/CD "the hard way", with many wasted hours editing `travis.yml`.
Such experiences "in the old days" (well, 2 years go) can easily lead one to falsely that new users should *also* learn and use tooling this way (cf the base vs tidyverse controversies).
So I might very well be wrong.

But I wonder whether the tradeoff is real.
And if the authors agree (that the tradeoff exists), it might be worthwhile adding a bit of a disclaimer to the package for newcomers.
I think R users who just quickly want to set up CI/CD should understand that using *tic* is a *choice*, and this choice should better be an informed one.
