### 2019-07-03 #305 @ldecicco-USGS
Note: Many help pages didn't have examples. I find examples really useful. Most of the functions that did have examples started with `dsl_init()`...and at the moment I'm writing this comment, I'm still confused what is happening (I fully expect that will change by the time I'm done with the review....in case I forget to come back here and edit my comments :)  )

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Full disclosure, I am not a `usethis` user...it might not be a bad idea for someone who uses and loves that package to review this package. I know it's popular, and I know people really like it, and I get the same vib from this package.

I couldn't figure out a way to use the package just by browsing the help files. That's not too bad...not all packages will work that way. It just happens to be what I do when I install a new package. So..., I started with the "Get Started" link:
https://ropenscilabs.github.io/tic/articles/tic.html

The first thing I did was run `tic::use_tic()`...and it was TERRIFYING. I have no idea what I was all authorizing. I was happily answering yes to overwriting local files, but then got really concerned about the github stuff going on (and those authorizations just pop up without a great deal of useful explaination from github what's happening...). I'd say to add more text to the description of the `use_tic` help file. I want to have the documentation open to know if I do indeed want to authorize github to do who-knows-what.

The thing that's frustrated me with CI stuff lately is that I have a great handle on Travis, but for some reason appveyor has been giving me all sorts of headaches. So, I like that theoretically I don't need to figure out appveyor. That being said, it's been empowering to me to understand the ins-and-outs of Travis.

My *FIRST* impression:

My tough question is....is this package a better use of development/maintenance resources compared to maintaining educational resources (blogs, tutorials, book?) on how to do it directly? Basically, weigh the risk/reward of having people "addicted" to this package vs creating their yml's from scratch. Reward: when R-users can't figure out their CI...they have the package maintainers of `tic` to ask why something isn't working - they can get back on track quicker, and have a more comprehensive CI setup from the get-go. Risk: when you all are off developing bigger and better packages, you might not have time to answer those users' questions or respond to a breaking API change in Travis or something like that (ie..users get frustrated with `tic` vs. getting frustrated with travis... 🙄 )

Also as an aside, my organization is strongly encouraging us to use a gitlab system, so now I need to figure out how to set up a ".gitlab-ci.yml" file. A blog like this:
https://blog.methodsconsultants.com/posts/developing-r-packages-with-usethis-and-gitlab-ci-part-iii/
has been incredibly useful to me. If I was already heavily invested in `tic`, I'm not sure if I'd have a seamless transition (maybe?).

My *SECOND* impression:

There are some great examples and use cases within the vignettes. I wonder if the "Get Started" page could be re-written to highlight those right away, to get users a reason to buy-in to the package concept (before running the `use_tic` function). I really liked the minimal example repos set up to show different use cases.

So for example...adding something right up front like:

You might already know that travis can:
* check your package
* run your tests
* figure out how well your code is tested

But did you know it could also:
* create the man files
* build the pkgdown site
* and push those back to your github site
* (something about drat....I don't quite follow that use-case)

If each of those bullets had a link to a place that described that concept in detail, coool.

My *last* impression (during the review):

I do think this is a useful package. If it gets published and well publicize, I think there could be many users that depend on it (perhaps heavily). As long as it's properly maintained, that's great. If there isn't long-term support for the package... as a user, I'd consider that a high-risk to me, and might not bother with it. Overall, 💯 , nice package, people will like it, and a few tweeks to the documentation might get MORE people to like it....but it's up to the developers to decide if that's a good thing.
