### 2019-07-04 #305 @annakrystalli
Thank you @ldecicco-USGS for your insightful review!

Over to you @krlmlr. I hope the reviews have provided some food for thought. They certainly have for me!
