### 2019-07-07 c66ffdd
- `do_package_checks()` gains `type` argument.
