### 2019-07-19 #305 @annakrystalli
Hi @krlmlr,

just touching base on this review, although I appreciate it is vacation season. Just let me know if for whatever reason you need more time to respond.
