### 2019-09-03 2bce433
bump R version requirement to 3.2.0 due to rlang
