+++ .travis.yml
@@ -51,18 +51,17 @@ jobs:
   include:
   - stage: test
     r: devel
-  - r: 3.2
   - stage: full
     r: release
     env:
     - BUILD_PKGDOWN=true
   - os: osx
     r: release
+  - r: 3.5
   - r: 3.4
   - r: 3.3
-  - r: 3.1
-  allow_failures:
-  - r: 3.1
+  - r: 3.2
+
 
 #matrix: Early abortion of failed builds
 matrix:
