+++ DESCRIPTION
@@ -20,7 +20,7 @@ License: GPL (>= 2)
 URL: https://github.com/ropenscilabs/tic
 BugReports: https://github.com/ropenscilabs/tic/issues
 Depends: 
-    R (>= 3.0.0)
+    R (>= 3.2.0)
 Imports: 
     backports,
     cli,
