### 2019-09-22 #305 @pat-s
Hi all,

since @krlmlr seems to be very busy right now, I'll take on replying to the reviewers comments.

> The first thing I did was run tic::use_tic()...and it was TERRIFYING. I have no idea what I was all authorizing. I was happily answering yes to overwriting local files, but then got really concerned about the github stuff going on (and those authorizations just pop up without a great deal of useful explaination from github what's happening...). I'd say to add more text to the description of the use_tic help file. I want to have the documentation open to know if I do indeed want to authorize github to do who-knows-what.

All steps that are being called by `use_tic` are outlined in the help page.
We could add more information on which of these steps trigger the browser windows. The function is already quite verbose and asks you before each step.
Certain tasks require opening browser windows. Would it help if we append this information also in the description of each step so that the user is not surprised by this action?

> My tough question is....is this package a better use of development/maintenance resources compared to maintaining educational resources (blogs, tutorials, book?) on how to do it directly? Basically, weigh the risk/reward of having people "addicted" to this package vs creating their yml's from scratch. Reward: when R-users can't figure out their CI...they have the package maintainers of tic to ask why something isn't working - they can get back on track quicker, and have a more comprehensive CI setup from the get-go. Risk: when you all are off developing bigger and better packages, you might not have time to answer those users' questions or respond to a breaking API change in Travis or something like that (ie..users get frustrated with tic vs. getting frustrated with travis... roll_eyes )

The ultimate goal is to be CI-agnostic. Specify what you want to do using R code as if you were working locally. Do not worry about the CI yml files anymore (in 90% of the time). _tic_ helps you also in writing R code by providing macros that do common things (install deps, run R cmd check, build and deploy pkgdown). You are always free to just write these steps in R code yourself.

> There are some great examples and use cases within the vignettes. I wonder if the "Get Started" page could be re-written to highlight those right away, to get users a reason to buy-in to the package concept (before running the use_tic function). I really liked the minimal example repos set up to show different use cases.

I think its hard to say what the best way is to promote the package/start with. Some people will always read first while others will always just execute code right away and "see what happens". It also depends on the level of experience with CI systems. Exp users will have a very different take on incorporating _tic_ than unexperienced ones.

> So for example...adding something right up front like:

> You might already know that travis can:
>
>     check your package
>     run your tests
>     figure out how well your code is tested
>
> But did you know it could also:
>
>     create the man files
>     build the pkgdown site
>     and push those back to your github site
>     (something about drat....I don't quite follow that use-case)

This looks like a good structure that we indeed could implement in the README to get people on board.

> I do think this is a useful package. If it gets published and well publicize, I think there could be many users that depend on it (perhaps heavily). As long as it's properly maintained, that's great. If there isn't long-term support for the package... as a user, I'd consider that a high-risk to me, and might not bother with it. Overall, 100 , nice package, people will like it, and a few tweeks to the documentation might get MORE people to like it....but it's up to the developers to decide if that's a good thing.

Thanks. Maintenance in this case means mainly the setup functions which mainly live in the _travis_ pkg. Everything else (macros, etc). is basically sugar for the user and input from the community is highly welcome.

#

Thanks for your review! I'll take the following actions items out of it:

- [ ] Provide more detailed information on the purpose and action of each step in `use_tic()`
- [ ] Attract more users by using better wording and promotion of the macro capabilities of _tic_ which enhance the "basic" Travis workflow
