### 2019-09-22 #305 @pat-s
# Review from @maxheld83

> I'm a bit skeptical as to whether it's a good idea to wrap yet another layer on top of the (usually) *.ymls used to express workflows on different CI services.
> These *.ymls, after all, are already abstractions over (bash?) scripts run by the services, and are meant as the UI for users.

Yes, they are already a layer on top of the CI system. However, these just install a basic toolchain for the respective language. Then again there needs to be a community-driven effort of maintaining that structure which is in fact [open-source](https://github.com/travis-ci/travis-build/blob/master/lib/travis/build/script/r.rb) but somewhat hidden and hard to understand for average people.
Speaking from the usage side, one has to learn the specific yaml options or discover them in the first place. Of course the same needs to be done for _tic_ if we would compare it with the macros offered by the package. However, the latter scale across CI systems whereas the former do not. And this is in the end the big advantage.

The second big advantage is the simplification of complex tasks.
These are made easy by allowing it to be written in R instead of dealing with a shell language (which most of the people are not so familiar with).

> There is an obvious appeal to abstracting over different providers, and for making this so smooth, but there are also some potential downsides:
>
> The complexity of managing CI/CD workflows grows considerably; now there's another abstraction layer which may have bugs or unexpected behavior.

True, this will be the case or at least a potential downside that thism pkg comes with.

> This will, among other things, require the authors to keep their package up to date with any upstream changes to the *.ymls by the CI/CD service providers.
> (A similar problem appears to be raised in #171)
> It's not obvious whether the abstraction (the "steps") actually do or will map elegantly to all providers now or in the future.
> For a counter-example, GitHub Actions doesn't know "steps."
> (GitHub Actions can be used for CI/CD, but it's still pretty new and weird, so this isn't a very relevant limitation right now -- and maybe it could be mapped).

This is also a valid concern. There is no way around having manual checks/updates for the compliancy of the template. However, we have example repositories running with the CI template files which should detect upstream changes early.
An upstream repo which is able to do automatic updates to child-repos containing the pure template file for a specific CI provider would be a nice addition - however I do not see a way how to implement this functionality easily right now (I do not know if this is possible at all while keeping the `git` setup behind this somewhat simple).

> It's not obvious that the interface of tic is more obvious or user-friendly than writing a *.yml, which isn't really that hard.

I would argue that this point is very subjective - a lot of users have huge problems understanding CI ymls. The ones who are not just deal with it frequently.

> On the one hand, for easy CI/CD workflows, one can usually just copy/paste a *.yml (plus some GitHub PAT action, admittedly).
> For very custom/complicated CI/CD workflows, on the other hand, learning the R6 class system of tic may be even more complicated than adding the logic of some providers *.yml.

The user is not bothered with R6 as all exported functionality are simple functions and not R6 classes to interact with.

> If wrapping *.ymls in another abstraction / DSL, one might wonder whether R is, in fact, the right language and level at which to do this.
> Arguably, writing service-agnostic workflows could be such a big undertaking (which the tic team seems to have largely covered, incredibly!), that it might be better organised under a "bigger tent" with a larger (non-R) community, that then simply gets wrapped for R convenience.

Correct. However, this project probably needs a lot more man power and serious funding behind it. And then again, R would probably not the first language to go for in that case. So I think there is no way but helping yourself here (yourself = R community).

>
> Absent such a bigger community, I hope that the authors will be able to cope with the maintenance load imposed by this (rapidly?) evolving/consolidating marketplace of tools.
> Alternatively, it would be interesting to see whether other communities (Python, Julia?) have also developed their own CI/CD abstraction/DSL.

Definitely, I have not checked on this yet.

> Lastly, one might wonder whether it's actually a good idea pedagogically, so to speak, to isolate R users from anything non-R (bash scripting, *.yml).
> I understand that many R users are not software developers and have limited patience and time for yet more tooling, but by wrapping ever more tools in R, we risk keeping (new) users relatively unprepared for the times when they will need to use outside (standard) tooling.

I agree but then again it is not our task to decide on this. This pkg is not just aiming at helping the "non-programming" R user but also for advanced people to simplify their workflow across providers.

> I've learned CI/CD "the hard way", with many wasted hours editing travis.yml.
> Such experiences "in the old days" (well, 2 years go) can easily lead one to falsely that new users should also learn and use tooling this way (cf the base vs tidyverse controversies).
> So I might very well be wrong.

On the other hand: What if Travis shuts its doors and the R com needs to move to another provider in the future? Then all the "Travis knowledge" is worth nothing. However, knowing a CI-agnostic system is key then.

> But I wonder whether the tradeoff is real.
> And if the authors agree (that the tradeoff exists), it might be worthwhile adding a bit of a disclaimer to the package for newcomers.
> I think R users who just quickly want to set up CI/CD should understand that using tic is a choice, and this choice should better be an informed one.

This is a good point to take-away.

---

Thanks @maxheld83 for your detailed thoughts on this! I'll share most of your concerns. However, most of them are also on a meta-level and hard to deal with. We will not know how the system develops and what might be the best approach to tackle CI workflows in the future. _tic_ offers one way (a CI-agnostic R DSL) and users need to decide whether they want to hop on that train or not.

Things I take away:

- [ ] Emphasize more in the README/Intro that _tic_ is a choice and that there are potential tradeoffs taking that route. People should know about the "layers of CI" and be advised to read the articles to gain a deeper understanding of what happens behind the scenes.
