### 2019-09-22 #305 @pat-s
@annakrystalli

I know it's been quite some time now but: Are there still open points for discussion from your review? Reading the comments history, I see that @krlmlr has already addressed your review in https://github.com/ropensci/software-review/issues/305#issuecomment-498207401.

As most of the comments from the other two reviews were on the meta site and somewhat hard to address/subjective, I "only" have the three action items listed at the bottom of my replies the reviews.

Do you still see any blocking issues apart from these points I listed that would block the transitioning to _ropensci_?

Again a **big thanks** to all reviewers (@annakrystalli, @ldecicco-USGS, @maxheld83 ) from my side and apologies for the late responses here.
