### 2019-09-22 f481f3a
disable tests on appveyor
