+++ tic.R
@@ -4,7 +4,8 @@ get_stage("install") %>%
 if (ci_on_travis()) {
   do_package_checks(error_on = if (getRversion() >= "3.2") "warning" else "error")
 } else if (ci_on_appveyor()) {
-  do_package_checks(args = "--no-tests", error_on = if (getRversion() >= "3.2") "warning" else "error")
+  do_package_checks(args = c("--no-tests", "--no-vignettes"),
+                    error_on = if (getRversion() >= "3.2") "warning" else "error")
 }
 
 if (ci_has_env("BUILD_PKGDOWN")) {
