+++ tic.R
@@ -5,6 +5,7 @@ if (ci_on_travis()) {
   do_package_checks(error_on = if (getRversion() >= "3.2") "warning" else "error")
 } else if (ci_on_appveyor()) {
   do_package_checks(args = c("--no-tests", "--no-vignettes"),
+                    build_args = "--no-build-vignettes",
                     error_on = if (getRversion() >= "3.2") "warning" else "error")
 }
 
