### 2019-09-23 #305 @annakrystalli
Hello @pat-s ! Great to hear from you, thanks for picking this up and for the detailed comments.

I'll allow @ldecicco-USGS & @maxheld83 to respond to them first. If they are both happy then the package will be accepted when the last points have been address. In general, it all sounds sensible to me but I'll defer to the reviewers for now.
