### 2019-09-23 #305 @ldecicco-USGS
Sounds great, I fully 👍 this package. Look for a feature request Github issue as I'm currently knee-deep in converting all my Travis jobs to .gitlab-ci.yml.... 😿
