### 2019-09-23 #305 @annakrystalli
Excellent! Thank you both! And sorry to hear that enforced migration is upon you @ldecicco-USGS.

So, as far as I'm concerned, please go ahead and make the last few changes you proposed @pat-s. When you're done, ping me here and I'll walk you through the last approval stages.

Thanks everyone! 👏
