### 2019-09-24 e82f48d
add basic circleci infrastructure
