### 2019-09-24 cf549d6
add check_dir arg to rcmdcheck
