### 2019-09-24 419d808
fix check_dir
