### 2019-09-25 d9e13af
use arch x64 and only binary packages
