### 2019-09-25 f637876
use arch x64 and only binary packages in own ci yml
