+++ appveyor.yml
@@ -36,6 +36,8 @@ on_finish: Rscript -e "tic::after_script()"
 #  - appveyor PushArtifact failure.zip
 
 environment:
+  PKGTYPE: binary
+  R_ARCH: x64
   GITHUB_PAT:
     secure: VXO22OHLkl4YhVIomSMwCZyOTx03Xf2WICaVng9xH7gISlAg8a+qrt1DtFtk8sK5
 
