### 2019-09-25 3075e07
use CRAN rcmdcheck, tidy DESCRIPTION
