### 2019-09-25 f8f2a80
readd vignettebuilder: knitr
