+++ DESCRIPTION
@@ -66,6 +66,8 @@ Suggests:
     testthat (>= 2.1.0),
     travis (>= 0.2.11.9001),
     usethis
+VignetteBuilder:
+  knitr
 Remotes: 
     ropenscilabs/travis
 ByteCompile: No
