+++ R/steps-install.R
@@ -18,7 +18,8 @@ InstallDeps <- R6Class(
         c(pkgType = private$type),
         remotes::install_deps(
           # https://github.com/r-lib/remotes/pull/386
-          dependencies = TRUE, repos = private$repos, build = FALSE
+          dependencies = TRUE, repos = private$repos, build = FALSE,
+          INSTALL_opts = "--no-multiarch"
         )
       )
     }
