+++ R/install.R
@@ -8,7 +8,10 @@ verify_install <- function(...) {
 
 verify_install_one <- function(pkg_name) {
   if (!package_installed(pkg_name)) {
+    withr::with_options(
+      c(pkgType = private$type),
     utils::install.packages(pkg_name, INSTALL_opts = "--no-multiarch")
+    )
     if (!package_installed(pkg_name)) {
       stopc(
         "Error installing package ", pkg_name, " or one of its dependencies."
