### 2019-09-25 c3e0fa8
use do.call() instead of lapply
