+++ R/install.R
@@ -11,7 +11,7 @@ verify_install <- function(pkg_names, pkgType = NULL) {
 }
 
 verify_install_one <- function(pkg_name, pkgType) {
-  if (!package_installed(pkg_name, pkgType)) {
+  if (!package_installed(pkg_name)) {
     withr::with_options(
       c(pkgType = pkgType),
     utils::install.packages(pkg_name, INSTALL_opts = "--no-multiarch")
