### 2019-09-25 9cb9e62
enforce default options of do_package_checks()
