### 2019-09-25 f6a92dc
enforce default options of do_package_checks()
