+++ R/install.R
@@ -6,7 +6,6 @@ verify_install <- function(pkg_names, pkgType = NULL) {
   if (is.null(pkgType)) {
     pkgType = update_type(pkgType)
   }
-  #do.call(verify_install, args = list(pkg_names, pkgType))
   lapply(pkg_names, function(x) verify_install_one(x, pkgType = pkgType))
 }
 
@@ -14,7 +13,7 @@ verify_install_one <- function(pkg_name, pkgType) {
   if (!package_installed(pkg_name)) {
     withr::with_options(
       c(pkgType = pkgType),
-    utils::install.packages(pkg_name, INSTALL_opts = "--no-multiarch")
+    utils::install.packages(pkg_name)
     )
     if (!package_installed(pkg_name)) {
       stopc(
