### 2019-09-25 6f01ccc
verbose git config
