+++ R/steps-git.R
@@ -270,9 +270,13 @@ DoPushDeploy <- R6Class(
         return(FALSE)
       }
 
+      # FIXME
+      print(git2r::config())
+
       message("Committing to ", git2r_attrib(private$git$get_repo(), "path"))
       new_commit <-
         git2r::commit(private$git$get_repo(), private$commit_message)$sha
+
 
       local <- git2r_head(private$git$get_repo())
       upstream <- git2r::branch_get_upstream(local)
