+++ appveyor.yml
@@ -36,6 +36,7 @@ on_finish: Rscript -e "tic::after_script()"
 #  - appveyor PushArtifact failure.zip
 
 environment:
+  R_REMOTES_STANDALONE: true # https://github.com/krlmlr/r-appveyor/issues/135
   GITHUB_PAT:
     secure: VXO22OHLkl4YhVIomSMwCZyOTx03Xf2WICaVng9xH7gISlAg8a+qrt1DtFtk8sK5
 
