### 2019-09-25 04b52d8
add argument check_dir to step_rcmdcheck
