+++ R/steps-rcmdcheck.R
@@ -6,7 +6,8 @@ RCMDcheck <- R6Class(
     initialize = function(warnings_are_errors = NULL, notes_are_errors = NULL,
                               args = c("--no-manual", "--as-cran"),
                               build_args = "--force", error_on = "warning",
-                              repos = repo_default(), timeout = Inf) {
+                              repos = repo_default(), timeout = Inf,
+                              check_dir = ".") {
       if (!is.null(notes_are_errors)) {
         warning_once(
           '`notes_are_errors` is deprecated, please use `error_on = "note"`'
@@ -29,6 +30,7 @@ RCMDcheck <- R6Class(
       private$error_on <- error_on
       private$repos <- repos
       private$timeout <- timeout
+      private$check_dir <- check_dir
 
       super$initialize()
     },
@@ -54,7 +56,8 @@ RCMDcheck <- R6Class(
           args = private$args, build_args = private$build_args,
           error_on = "never",
           repos = private$repos,
-          timeout = private$timeout
+          timeout = private$timeout,
+          check_dir = private$check_dir
         )
       )
 
@@ -87,7 +90,8 @@ RCMDcheck <- R6Class(
     build_args = NULL,
     error_on = NULL,
     repos = NULL,
-    timeout = NULL
+    timeout = NULL,
+    check_dir = NULL
   )
 )
 
@@ -118,6 +122,8 @@ RCMDcheck <- R6Class(
 #' @param timeout `[numeric]`\cr
 #'   Passed to `rcmdcheck::rcmdcheck()`, default:
 #'   `Inf`.
+#' @param check_dir `[character]` \cr Path specifying the directory for R CMD
+#'   check. Defaults to project root for easy upload of artifacts.
 #' @export
 #' @examples
 #' dsl_init()
@@ -129,7 +135,7 @@ RCMDcheck <- R6Class(
 step_rcmdcheck <- function(...,
                            warnings_are_errors = NULL, notes_are_errors = NULL,
                            args = NULL, build_args = NULL, error_on = "warning",
-                           repos = repo_default(), timeout = Inf) {
+                           repos = repo_default(), timeout = Inf, check_dir = ".") {
 
   #' @param build_args `[character]`\cr
   #'   Passed to `rcmdcheck::rcmdcheck()`.\cr
@@ -167,7 +173,8 @@ step_rcmdcheck <- function(...,
     build_args = build_args,
     error_on = error_on,
     repos = repos,
-    timeout = timeout
+    timeout = timeout,
+    check_dir = check_dir
   )
 }
 
