+++ R/macro-package-checks.R
@@ -36,7 +36,7 @@ do_package_checks <- function(...,
                               build_args = NULL,
                               error_on = "warning",
                               repos = repo_default(), timeout = Inf,
-                              type = NULL) {
+                              type = NULL, check_dir = ".") {
   #' @description
   #' 1. [step_install_deps()] in the `"install"` stage, using the
   #'    `repos` argument.
@@ -57,7 +57,8 @@ do_package_checks <- function(...,
         build_args = !!enquo(build_args),
         error_on = !!enquo(error_on),
         repos = !!enquo(repos),
-        timeout = !!enquo(timeout)
+        timeout = !!enquo(timeout),
+        check_dir = !!enquo(check_dir)
       )
     )
 
