### 2019-09-25 5cb1dce
move check_dir example to own PR
