+++ man/step_rcmdcheck.Rd
@@ -6,8 +6,7 @@
 \usage{
 step_rcmdcheck(..., warnings_are_errors = NULL,
   notes_are_errors = NULL, args = NULL, build_args = NULL,
-  error_on = "warning", repos = repo_default(), timeout = Inf,
-  check_dir = NULL)
+  error_on = "warning", repos = repo_default(), timeout = Inf)
 }
 \arguments{
 \item{...}{Ignored, used to enforce naming of arguments.}
