### 2019-10-01 160b35c
isTRUE(travis::uses_github()) always FALSE on R < 3.5
Although the output of travis::uses_github() is of type atomic and logical, when it is TRUE is does not pass test isTRUE because of its attributes, at least on R 3.4.4 with travis 0.2.11.9001. It is said in R base 3.4.4 documentation :
"isTRUE(x) is an abbreviation of identical(TRUE, x), and so is true if and only if x is a length-one logical vector whose only element is TRUE and which has no attributes (not even names)."
thus giving FALSE when there are some attributes, see https://www.r-bloggers.com/working-with-istrue/ for an example.
It
[…]
