### 2019-10-02 f257ebf
add verify_update_one()
