### 2019-10-08 abf2bee
Use remotes::install_cran(upgrade = TRUE) to install packages
