+++ .circleci/config.yml
@@ -1,0 +1,53 @@
+version: 2
+jobs:
+  build:
+    docker:
+      - image: rocker/verse
+    steps:
+      - checkout
+
+      # create a unique env var for the cache. Unfortunately normal env vars
+      # are not picked up by the cache, therefore this workaround is needed.
+      # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
+      - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
+      - restore_cache:
+
+      # install deps and check pkg ---------------------------------------------
+      - run:
+          name: Install package dependencies
+          command: |
+            echo -e 'options(Ncpus = 6, repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
+            R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
+            R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
+            R -q -e 'remotes::install_github("ropenscilabs/tic", upgrade = "always"); print(tic::dsl_load()); tic::prepare_all_stages()'
+            R -q -e 'tic::before_install()'
+            R -q -e 'tic::install()'
+      - run:
+          name: Check package
+          no_output_timeout: 60m
+          command: |
+            R -q -e 'tic::before_script()'
+            R -q -e 'tic::script()'
+
+      # save R pkg cache -------------------------------------------------------
+      - save_cache:
+          paths:
+            - /usr/local/lib/R/site-library
+
+workflows:
+  version: 2
+  build-and-deploy:
+    jobs:
+      - build
+# CRON job daily at 3 am in the morning
+# - runs the "build" job on the master branch and builds package cache
+  nightly:
+    triggers:
+      - schedule:
+          cron: "0 3 * * *"
+          filters:
+            branches:
+              only:
+                - master
+    jobs:
+      - build
