+++ .circleci/config.yml
@@ -11,6 +11,7 @@ jobs:
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
       - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
       - restore_cache:
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
 
       # install deps and check pkg ---------------------------------------------
       - run:
@@ -31,16 +32,17 @@ jobs:
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
           paths:
             - /usr/local/lib/R/site-library
 
 workflows:
   version: 2
-  build-and-deploy:
+  build:
     jobs:
       - build
 # CRON job daily at 3 am in the morning
-# - runs the "build" job on the master branch and builds package cache
+# - runs the "build" job on the master branch and buils the package cache
   nightly:
     triggers:
       - schedule:
