### 2019-10-09 72148c3
check if we need circle in use_tic.R
