+++ R/steps-git.R
@@ -462,12 +462,14 @@ PushDeploy <- R6Class(
 #'
 #' @export
 #' @examples
+#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script") %>%
 #'   add_step(step_push_deploy(commit_paths = c("NAMESPACE", "man")))
 #'
 #' dsl_get()
+#' }
 step_push_deploy <- function(path = ".", branch = NULL,
                              remote_url = NULL,
                              commit_message = NULL, commit_paths = ".") {
