### 2019-10-09 3b1650c
- added support for Circle CI (#177)
* add basic circleci infrastructure

* add check_dir arg to rcmdcheck

* fix check_dir

* use new arg in macro

* document

* verbose git config

* verbosity

* update circleci template

* update README

* move check_dir example to own PR

* fix ci()

* clean

* use_circle_yml() for use_tic()

* clean

* no deployment on circle

* add circleci yaml

* fix circle yaml

* check if we need circle in use_tic.R

* dont run dsl_load() examples due to circle issues

* add badge

* more dontrun

* more dontrun

* more dontrun
