+++ DESCRIPTION
@@ -79,6 +79,7 @@ Collate:
     'ci.R'
     'appveyor.R'
     'base64.R'
+    'circleci.R'
     'dsl-storage.R'
     'dsl.R'
     'git2r.R'
