### 2019-10-09 d0cb15f
check_dir = NULL
