### 2019-10-09 c55256c
Deploy from Travis build 1180 [ci skip]
Build URL: https://travis-ci.org/ropenscilabs/tic/builds/595576303
Commit: 8d7c6255677c8df0912392c84220401ab8cb9bd6
