+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: '2.2'
-pkgdown: 1.3.0
+pkgdown: 1.4.1
 pkgdown_sha: ~
 articles:
   advanced: advanced.html
