### 2019-10-09 043f9d5
- add argument "check_dir" to step_rcmdcheck (#179)
* add argument check_dir to step_rcmdcheck

* check_dir = NULL
