+++ man/do_package_checks.Rd
@@ -7,7 +7,7 @@
 do_package_checks(..., codecov = !ci_is_interactive(),
   warnings_are_errors = NULL, notes_are_errors = NULL, args = NULL,
   build_args = NULL, error_on = "warning", repos = repo_default(),
-  timeout = Inf, type = NULL)
+  timeout = Inf, type = NULL, check_dir = NULL)
 }
 \arguments{
 \item{...}{Ignored, used to enforce naming of arguments.}
@@ -52,6 +52,9 @@ Passed to \code{rcmdcheck::rcmdcheck()}, default:
 \item{type}{Passed on to \code{\link[=install.packages]{install.packages()}}. The default avoids
 installation from source on Windows and macOS by passing
 \code{\link{.Platform}$pkgType}.}
+
+\item{check_dir}{\code{[character]} \cr Path specifying the directory for R CMD
+check. Defaults to project root for easy upload of artifacts.}
 }
 \description{
 \code{do_package_checks()} adds default steps related to package checks
