### 2019-10-14 a7b5574
fix ci_on_circle()
