+++ DESCRIPTION
@@ -40,7 +40,7 @@ BugReports: https://github.com/ropenscilabs/tic/issues
 Depends: 
     R (>= 3.2.0)
 Imports: 
-    backports,
+    backports (>= 1.1.5),
     cli,
     crayon,
     magrittr,
