### 2019-10-16 cf582ad
fix ci_get_slug() for circle
