+++ inst/templates/dot-travis-macos-matrix.yml
@@ -1,0 +1,44 @@
+# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+# Default configuration for use with tic package
+# tic documentation to get started: https://ropenscilabs.github.io/tic/
+# Usually you shouldn't need to change the first part of the file
+
+# DO NOT CHANGE THE CODE BELOW
+before_install:
+  - R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
+  - R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
+  - R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
+  - R -q -e 'tic::before_install()'
+install: R -q -e 'tic::install()'
+after_install: R -q -e 'tic::after_install()'
+before_script: R -q -e 'tic::before_script()'
+script: R -q -e 'tic::script()'
+after_success: R -q -e 'tic::after_success()'
+after_failure: R -q -e 'tic::after_failure()'
+# DO NOT CHANGE THE CODE ABOVE
+
+# Custom parts:
+
+# Header
+language: r
+os: osx
+cache:
+  - packages
+  - $HOME/.ccache
+latex: false
+
+matrix:
+  allow_failures:
+  - r: oldrel
+  include:
+  - r: devel
+  - r: oldrel
+  - r: release
+
+#env
+env:
+  global:
+  - MAKEFLAGS="-j 2"
+
+#services
+services:
