+++ inst/templates/dot-travis-macos-deploy.yml
@@ -1,7 +1,5 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 # Default configuration for use with tic package
 # tic documentation to get started: https://ropenscilabs.github.io/tic/
-# Usually you shouldn't need to change the first part of the file
 
 # DO NOT CHANGE THE CODE BELOW
 before_install:
@@ -25,20 +23,14 @@ after_deploy: R -q -e 'tic::after_deploy()'
 after_script: R -q -e 'tic::after_script()'
 # DO NOT CHANGE THE CODE ABOVE
 
-# Custom parts:
-
-# Header
 language: r
 os: osx
 cache:
   - packages
   - $HOME/.ccache
+  - $HOME/Library/Caches/Homebrew
 latex: false
 
-#env
 env:
   global:
   - MAKEFLAGS="-j 2"
-
-#services
-services:
