+++ inst/templates/dot-travis-linux.yml
@@ -1,7 +1,5 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 # Default configuration for use with tic package
 # tic documentation to get started: https://ropenscilabs.github.io/tic/
-# Usually you shouldn't need to change the first part of the file
 
 # DO NOT CHANGE THE CODE BELOW
 before_install:
@@ -19,9 +17,6 @@ before_deploy: R -q -e 'tic::before_deploy()'
 after_script: R -q -e 'tic::after_script()'
 # DO NOT CHANGE THE CODE ABOVE
 
-# Custom parts:
-
-# Header
 language: r
 dist: bionic
 cache:
@@ -29,10 +24,6 @@ cache:
   - $HOME/.ccache
 latex: false
 
-#env
 env:
   global:
   - MAKEFLAGS="-j 2"
-
-#services
-services:
