+++ inst/templates/circle.yml
@@ -11,7 +11,7 @@ jobs:
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
       - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
+         key: R-package-library-checksum "/tmp/_tmp_file"
 
       # install deps and check pkg ---------------------------------------------
       - run:
@@ -32,7 +32,7 @@ jobs:
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
+          key: R-package-library-checksum "/tmp/_tmp_file"
           paths:
             - /usr/local/lib/R/site-library
 
