+++ inst/templates/circle-deploy.yml
@@ -1,17 +1,19 @@
+# tic documentation to get started: https://ropenscilabs.github.io/tic/
+
 version: 2
 jobs:
-  build:
+  "r-3.6.1":
     docker:
-      - image: rocker/verse
+      - image: rocker/verse:3.6.1
     steps:
       - checkout
 
       # create a unique env var for the cache. Unfortunately normal env vars
       # are not picked up by the cache, therefore this workaround is needed.
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
-      - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
+      - run: echo "$(date -d "$D" '+%d')3.6.1" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
+          key: R-package-library-checksum "/tmp/_tmp_file"
 
       # install deps and check pkg ---------------------------------------------
       - run:
@@ -32,7 +34,7 @@ jobs:
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
+          key: R-package-library-checksum "/tmp/_tmp_file"
           paths:
             - /usr/local/lib/R/site-library
   deploy:
@@ -44,12 +46,7 @@ jobs:
       - checkout
       - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
-      # Add your own SSH key fingerprint here, see https://circleci.com/docs/2.0/gh-bb-integration/#creating-a-github-deploy-key
-      # and https://discuss.circleci.com/t/adding-ssh-keys-fails/7747/24 on how to create a valid key
-      #- add_ssh_keys:
-      #    fingerprints:
-      #      - "15:9c:45:50:dc:94:87:6f:a9:ca:e1:fc:6a:b9:7e:62"
+          key: R-package-library-checksum "/tmp/_tmp_file"
       - run:
           name: Deploy Over SSH
           command: |

[…]
