### 2019-10-24 6bef175
first version of new use_tic()
