+++ R/use_tic.R
@@ -1,13 +1,16 @@
-# This code can only run interactively
-# nocov start
-
-#' Initialize CI testing using tic
+#' @title Initialize CI testing using tic
+#' @description Prepares a repo for building and deploying supported by
+#' \pkg{tic}.
 #'
-#' Prepares a repo for building and deploying supported by \pkg{tic}.
+#' @details
 #'
-#' @param quiet `[flag]`\cr
-#'   Less verbose output? Default: `FALSE`.
+#' 1. Query information which CI providers should be used
+#' 1. Setup permissions for providers selected for deployment
+#' 1. Create YAML files for selected providers
+#' 1. Create a default `tic.R` file depending on the repo type
+#'    (package, website, bookdown, ...)
 #'
+#' @param quiet `[flag]`\cr Less verbose output? Default: `FALSE`.
 #' @export
 use_tic <- function(quiet = FALSE) {
   cli_alert("Welcome to {.pkg tic}!")
@@ -22,12 +25,16 @@ use_tic <- function(quiet = FALSE) {
 
   cli_par()
   cli_text(c("There is no need to use all of them.",
-           " You can choose which one(s) you want to use and on which systems you want to build on.")
+             " You can choose which one(s) you want to use,",
+             " whether you want to deploy (i.e. push from builds)",
+             " and if you want to test on multiple R versions.")
   )
   cli_end()
 
   cli_text("We recommend the following setup:")
   cli_ul(c("Appveyor: Windows", "Circle CI: Linux", "Travis CI: macOS"))
+  cli_par()
+  cli_end()
 
   if (yesno("Ready to get started?")) {
     return(invisible())
@@ -35,6 +42,8 @@ use_tic <- function(quiet = FALSE) {
 
   cli_h1("Choosing your setup.")
   cli_text("We'll ask you a few yes/no questions to gather your preferences.")
+  cli_par()
+  cli_end()
 
   windows = menu(c("Yes", "No"), title = "Do you want to build on Windows (= Appveyor)?")
   mac = menu(c("Yes", "No"), title = "Do you want to build on macOS (= Travis CI)?")
@@ -42,42 +51,49 @@ use_tic <- function(quiet = FALSE) {
   linux = menu(c("Circle CI", "Travis CI", "None", "All"),
                title = "Which provider do you want to use for Linux builds?")
 
-  deploy = menu(c("Circle CI", "Travis CI", "None", "All"),
+  deploy = menu(c("Circle CI", "Travis CI", "No", "All"),
                 title = "Do you want to deploy (i.e. push from the CI build to your repo) on certain providers? If yes, which ones?")
 
-  matrix = menu(c("Circle CI", "Travis CI", "Appveyor", "None", "All"),
+  matrix = menu(c("Circle CI", "Travis CI", "Appveyor", "No", "All"),
                 title = "Do you want to build on multiple R versions? (i.e. R-devel, R-release, R-oldrelease). If yes, on which platform(s)?")
 
   cli_h1("Setting up the CI providers.")
 
-  cli_text("Now we are getting the selected CI providers ready for deployment.",
-           "This requires some interaction with their API and you may need to create an API token.")
+  cli_text("Next we are getting the selected CI providers ready for deployment.",
+           " This requires some interaction with their API and you may need to create an API token.")
+  cli_par()
+  cli_end()
 
+  browser()
   # init deploy ----------------------------------------------------------------
 
   if (deploy == 1) {
+    cat(boxx("Circle CI", border_style = "double"))
     check_circle_pkg()
[…]
