+++ DESCRIPTION
@@ -32,8 +32,7 @@ Description: Provides a way to describe common build and
     deployment workflows for R-based projects: packages, websites (e.g.
     blogdown, pkgdown), or data processing (e.g. research compendia).  The
     recipe is described independent of the continuous integration tool
-    used for processing the workflow (e.g.  'Travis CI' or
-    'AppVeyor').
+    used for processing the workflow (e.g.  'Travis CI' or 'AppVeyor').
 License: GPL (>= 2)
 URL: https://github.com/ropenscilabs/tic
 BugReports: https://github.com/ropenscilabs/tic/issues
@@ -54,6 +53,7 @@ Suggests:
     base64enc,
     bookdown,
     callr,
+    circle,
     covr,
     devtools,
     drat,
@@ -67,8 +67,9 @@ Suggests:
     travis (>= 0.2.11.9001),
     usethis
 VignetteBuilder:
-  knitr
+    knitr
 Remotes: 
+    pat-s/circle,
     ropenscilabs/travis
 ByteCompile: No
 Encoding: UTF-8
@@ -107,5 +108,6 @@ Collate:
     'steps-write-text-file.R'
     'tic-package.R'
     'travis.R'
+    'use-yaml.R'
     'use_tic.R'
     'utils.R'
