+++ vignettes/tic.Rmd
@@ -29,167 +29,89 @@ See the ["Features" article](features.html) for a comparison with other solution
 # Initialization
 
 The easiest way to use *tic* for CI services is to call `tic::use_tic()`.
-This will create templates for both "Travis CI" and "Appveyor" and initialize all the required authentication steps for deployment.
-During the process, browser pages will open up, ensuring that all permissions are set correctly and all apps are authorized.
-Also a Personal Access Token (PAT) will be created on GitHub and stored  as an encrypted variable on Travis.
-This avoids hitting rate limitations when accessing the GitHub API, see [Locke Data's blog post about GitHUB PATs](https://itsalocke.com/blog/using-travis-make-sure-you-use-a-github-pat/) for more details.
-The PAT creation cannot be fully automated as some input from the user on GitHub is required when creating the PAT.
+This will initialize a setup wizard which will guide you through all possibilities of the offered CI providers by _tic_.
 
-First, authentication with Travis (a browser window opens) and GitHub is initiated (another browser window opens).
-Also, the repo is activated on Travis.
+Several yes/no questions need to be answered. 
+Based on the replies we'll select specific templates.
+
+Besides the question which CI system you want to use, you'll be asked if you want to 
+
+- deploy from builds to Github (e.g. if you are building a pkgdown site)
+
+- test your package on multiple R versions
 
 ```r
 tic::use_tic()
-┌───────────────────────┐
-│                       │
-│   Welcome to `tic`!   │
-│                       │
-└───────────────────────┘
-ℹ This wizard will set all the required tokens and files
-  on Travis CI and GitHub. Lets get started!
-✔ Setting active project to '/home/pjs/git/test2'
-ℹ Using active project '/home/pjs/git/test2'
-┌──────────────────────────────────────────────────────┐
-│                                                      │
-│   Step #1: We check if a GitHub repository exists.   │
-│                                                      │
-└──────────────────────────────────────────────────────┘
-❯  Authenticating to GitHub.
-Waiting for authentication in browser...
-Press Esc/Ctrl + C to abort
-Authentication complete.
-✔ Authentication successful.
-❯  Authenticating to GitHub.
-Waiting for authentication in browser...
-Press Esc/Ctrl + C to abort
-Authentication complete.
-✔ Authentication successful.
-✔ GitHub repo exists.
-┌─────────────────────────────────────────────────────┐
-│                                                     │
-│   Step #2: We check if Travis is already enabled.   │
-│                                                     │
-└─────────────────────────────────────────────────────┘
-❯  Authenticating to GitHub.
-Waiting for authentication in browser...
-Press Esc/Ctrl + C to abort
-Authentication complete.
-✔ Authentication successful.
-❯  Authenticating to GitHub.
-❯  Authenticating to GitHub.
-Waiting for authentication in browser...
-Press Esc/Ctrl + C to abort
-Authentication complete.
-✔ Authentication successful.
-Finished activating repo pat-s/test2 on Travis CI.
-```
 
-Next, you will be asked if the current `.travis.yml` should be replaced with the template from *tic* (if one exists).
-Otherwise, the template is added without asking.
+→ Welcome to → tic!
+This wizard will guide you through the setup process for getting started with various CI
+providers.
 
-```r
-┌────────────────────────────────────────────────┐
-│                                                │
-│   Step #3: We create new files                 │
-│   `.travis.yml`, `appveyor.yml` and `tic.R`.   │
[…]
