### 2019-10-31 81a1d88
finish use_tic()
