+++ R/use-yaml.R
@@ -46,6 +46,7 @@ use_circle_yml <- function(type) {
   } else if (type == "linux-deploy-matrix") {
     template = "circle-deploy-matrix.yml"
   }
+  dir.create(".circleci")
   use_tic_template(
     template,
     save_as = ".circleci/config.yml",
