+++ man/tic-package.Rd
@@ -10,8 +10,7 @@ Provides a way to describe common build and
     deployment workflows for R-based projects: packages, websites (e.g.
     blogdown, pkgdown), or data processing (e.g. research compendia).  The
     recipe is described independent of the continuous integration tool
-    used for processing the workflow (e.g.  'Travis CI' or
-    'AppVeyor').
+    used for processing the workflow (e.g.  'Travis CI' or 'AppVeyor').
 }
 \details{
 The \code{\link[=use_tic]{use_tic()}} function prepares a code repository for use with
