### 2019-10-31 5e4482a
remove after_install step
