+++ inst/templates/dot-travis-macos-deploy-matrix.yml
@@ -8,7 +8,6 @@ before_install:
   - R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
   - R -q -e 'tic::before_install()'
 install: R -q -e 'tic::install()'
-after_install: R -q -e 'tic::after_install()'
 before_script: R -q -e 'tic::before_script()'
 script: R -q -e 'tic::script()'
 after_success: R -q -e 'tic::after_success()'
