+++ NAMESPACE
@@ -67,8 +67,11 @@ export(step_setup_push_deploy)
 export(step_setup_ssh)
 export(step_test_ssh)
 export(step_write_text_file)
+export(use_appveyor_yml)
+export(use_circle_yml)
 export(use_tic)
 export(use_travis_deploy)
+export(use_travis_yml)
 import(backports)
 import(cli)
 import(rlang)
