### 2019-10-31 #305 @pat-s
Hi everyone,

@krlmlr and I decided to make the workhorse function `use_tic()` act smarter and a lot more interactive as before. This should give the user a better idea of what is happening.

Also the user can soon choose between three platforms (circle ci, travis ci, appveyor) and make choices whether to
- (1) deploy
- (2) build on multiple R versions
- (3) build on Travis or macOS on those platforms.

Depending on the choices different actions will be taken (e.g. setting up deployment keys for specific providers) and specific YAML templates will be selected.

We are now also using the _cli_ package for the interactive messages.

This, in combination with ongoing work in the _travis_ package (with a full refactoring of the auth process and no more browser tab popups) and the new [_circle_](https://github.com/pat-s/circle) package (for interfacing the Circle CI API) should give the whole project a new kickstart and make it a lot easier for users.

We will meet in person at the end of Nov to finalize things. Until then, we hope that https://github.com/ropenscilabs/tic/pull/193 is finished and all the other changes are done so that we finally can go to ropensci and CRAN.

We would appreciate user feedback for https://github.com/ropenscilabs/tic/pull/193 if you have some minutes to spare 🙏
