### 2019-11-03 30dfeeb
test conditional exec of step_setup_ssh()
