### 2019-11-03 5000aba
fix R CMD check
