### 2019-11-03 fbb0346
ci_can_push() always true
