### 2019-11-03 7f20daa
explicit condition
