### 2019-11-03 c98d652
Update support for Circle CI deployment (#195)
* test conditional exec of step_setup_ssh()

* fix R CMD check

* require circle

* require circle

* install circle

* ci_can_push() always true

* explicit condition
