### 2019-11-18 9a5f965
Need dev rlang for formatting of conditions
