### 2019-11-29 afd2c0f
update tic.Rmd
