+++ vignettes/tic.Rmd
@@ -26,7 +26,7 @@ If you are unfamiliar with the term Continuous Integration (CI), we highly recom
 See the ["Features" article](features.html) for a comparison with other solutions for Continuous Integration for R.
 
 
-# Initialization
+# Initialization/Setup
 
 The easiest way to use *tic* for CI services is to call `tic::use_tic()`.
 This will initialize a setup wizard which will guide you through all possibilities of the offered CI providers by _tic_.
@@ -107,7 +107,7 @@ Last, we'll add a `tic.R` file to the project root.
 The deployment setup functionality comes from the _travis_ and _circle_ packages.
 See the ["tic & travis" article](tic-travis.html) for more detailed information on how *tic* and *travis* work together.
 
-# Explanation of the basic template
+# Explanation of the basic `tic.R` template
 
 After having called `tic::use_tic()` you will find a `.travis.yml`, `.circleci/config.yml`,  `appveyor.yml` and a `tic.R` file in your repo.
 Usually you do not need to touch the YAML files anymore.
@@ -124,16 +124,24 @@ if (ci_on_travis()) {
 }
 ```
 
-Let's break down what happens here:
+## Macros
+
+_tic_ builds on the "macro" idea.
+Macros are essentially wrappers of a sequence of steps for often used tasks on the CI system:
+
+- Checking a package (R CMD check)
+- Building and deploying a pkgdown site
+- Build and deploy a bookdown project
 
 ```{r eval = FALSE}
 do_package_checks()
 ```
 
-`do_package_checks()` Is a macro which adds essential steps to various stages of a CI run. 
-Essentially, it adds `step_rcmdcheck()` to the "script" stage. 
-This step performs the check of a R package.
+`do_package_checks()` adds essential steps to various stages of a CI run. 
+Most importantly, it adds `step_rcmdcheck()` to the "script" stage. 
+This step performs the check of an R package.
 Afterwards, the code coverage is being checked using `covr::codecov()`.
+See `?do_package_checks()` for more information.
 
 ```{r}
 # step_install_deps() in the "install" stage, using the repos argument.
@@ -143,7 +151,7 @@ Afterwards, the code coverage is being checked using `covr::codecov()`.
 # A call to covr::codecov() in the "after_success" stage (only if the codecov flag is set)
 ```
 
-The other macro in the default template is `do_pkdown()`.
+The other macro in the default template is `do_pkgdown()`.
 
 ```{r eval = FALSE}
 if (ci_on_travis()) {
@@ -165,13 +173,10 @@ if (ci_on_travis()) {
 # step_do_push_deploy() in the "deploy" stage.
 ```
 
-In a nutshell: All dependencies are being installed, the deployment setup is initialized, the pkgdown site is build and deployed.
+By default this currently happens only on Travis CI, because `ci_on_travis()` is used as a condition.
+Why do we do this? Building the pkgdown site on multiple CI services has no added benefit and might even cause problems due to race conditions during deployment.
 
-This happens only on Travis CI, because `ci_on_travis()` evaluates to `TRUE` only on that platform.
-This is also the default in the [template](https://github.com/ropenscilabs/tic/blob/e7649fc32cd5071f62289974af7f38bd36f90cd3/inst/templates/package/tic.R) for packages.
-Building the pkgdown site on multiple CI services might cause problems and increases build time for no reason.
-
-# Examples projects
+# Example projects
 
 *tic* can be used for various R projects:
 
@@ -184,6 +189,8 @@ Building the pkgdown site on multiple CI services might cause problems and incre
[…]
