### 2019-11-30 e1013bb
update vignettes
