+++ vignettes/advanced.Rmd
@@ -23,51 +23,50 @@ Possible conditionals are
 
 - Environment variables of the build (queried by `ci_is_env()`, `ci_has_env()` or `ci_get_env()`)
 - R Version of the current build (`getRversion()`)
-- other features of the build (e.g. branch name, `ci_get_branch()`)
+- other features of the build (e.g. branch name via `ci_get_branch()`)
 
-The most common use cases are the testing on multiple R versions and the restriction of certain tasks that should only be executed once (e.g. the [deployment of a _pkgdown_ site](deployment.html#pkgdown-deployment)).
-If you test on multiple CI services you may also want to run certain tasks on only one of them.
+Most common tasks are the testing on multiple R versions and the restriction of certain tasks that should only be executed once (e.g. the [deployment of a _pkgdown_ site](deployment.html#pkgdown-deployment)).
 
 ### Conditional execution: Use cases
 
-The following shows you some example code blocks to condition certain stages and their respective steps on 
+The following shows some example code blocks to condition certain stages and their respective steps on 
 
 - the R Version (using an environment variable)
-- the CI service
-- a specific branch
 
 ```{r eval = FALSE}
+# this env var already exists by default on Travis
+# for other systems you need to set a custom env var in the CI YAML
 if (ci_is_env("TRAVIS_R_VERSION_STRING", "release")) {
   get_stage("after_success") %>%
     add_step(step_...())
 }
 ```
 
+- the CI service
+
 ```{r eval = FALSE}
 if (ci_on_travis()) {
-  get_stage("<a stage>") %>%
+  get_stage("<stage>") %>%
     add_step(step_...())
 }
 ```
 
+- a specific branch
+
 ```{r eval = FALSE}
 if (ci_get_branch() == "master") {
-  get_stage("<a stage>") %>%
+  get_stage("<stage>") %>%
     add_step(step_...())
 }
 ```
 
 ## Installation of packages
 
-Usually required packages are installed based on the DESCRIPTION file contents of a package.
+Required packages are installed based on the Depends/Imports/Suggests fields of the DESCRIPTION file.
 You should only use the following steps in extraordinary circumstances.
-
-Also the manual installation of GitHub or CRAN packages should only be used in non-interactive CI runs. 
-This can be checked with the condition `ci_is_interactive()`.
+An example can be the use of a package in the REAMDE.Rmd file which is not listed within the package's DESCRIPTION file.
 
 ### GitHub packages
-
-GitHub packages should be installed in the `install` stage:
 
 ```{r eval = FALSE}
 get_stage("install") %>%
@@ -80,11 +79,7 @@ Note that the underlying `remotes::install_github()` is vectorized for the `repo
 add_step(step_install_github(c("r-lib/rcmdcheck", "r-lib/usethis")))
 ```
 
-
 ### CRAN packages
-
-Sometimes you need to install CRAN packages that are not specified in the DESCRIPTION file of your package.
-To install the package (only if it does not exist in the cache already), run the following:
 
 ```{r eval = FALSE}
 get_stage("install") %>%
@@ -94,9 +89,9 @@ get_stage("install") %>%
 ## CI Meta-Information
 
 The `ci()` function and its friends (`ci_*`) hold valuable information about the CI system. 
-It can be used to query information that can be again be utilized for conditioning stages or steps.
+They can be used to query information that can be again be utilized for conditioning stages or steps.
 
-For example, the user may wish to only deploy on Travis by using `ci_on_travis()`:
+For example, the user may wish to only deploy on Travis CI by using `ci_on_travis()`:
 
 ```{r eval = FALSE}
 if (ci_on_travis()) {
@@ -108,6 +103,7 @@ if (ci_on_travis()) {
 }
 ```
 
+By using the code above, the specified steps will only be executed on Travis CI.
[…]
