+++ inst/templates/travis-no-deploy.yml
@@ -1,7 +1,7 @@
 # Default configuration for use with tic package
-# tic documentation to get started: https://ropenscilabs.github.io/tic/
+# tic documentation to get started: https://docs.ropensci.org/tic/dev/
 
-# DO NOT CHANGE THE CODE BELOW
+# DO NOT CHANGE THE CODE BELOW -------------------------------------------------
 before_install:
   - R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
   - R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
@@ -12,31 +12,7 @@ before_script: R -q -e 'tic::before_script()'
 script: R -q -e 'tic::script()'
 after_success: R -q -e 'tic::after_success()'
 after_failure: R -q -e 'tic::after_failure()'
-before_deploy: R -q -e 'tic::before_deploy()'
-deploy:
-  provider: script
-  script: R -q -e 'tic::deploy()'
-  on:
-    all_branches: true
-after_deploy: R -q -e 'tic::after_deploy()'
 after_script: R -q -e 'tic::after_script()'
-# DO NOT CHANGE THE CODE ABOVE
+# DO NOT CHANGE THE CODE ABOVE -------------------------------------------------
 
-language: r
-dist: bionic
-cache:
-  - packages
-  - $HOME/.ccache
-latex: false
 
-env:
-  global:
-  - MAKEFLAGS="-j 2"
-
-matrix:
-  allow_failures:
-  - r: oldrel
-  include:
-  - r: devel
-  - r: oldrel
-  - r: release
