+++ inst/templates/travis-macos.yml
@@ -1,30 +1,5 @@
 # Default configuration for use with tic package
 # tic documentation to get started: https://ropenscilabs.github.io/tic/
 
-# DO NOT CHANGE THE CODE BELOW
-before_install:
-  - R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
-  - R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
-  - R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
-  - R -q -e 'tic::before_install()'
-install: R -q -e 'tic::install()'
-before_script: R -q -e 'tic::before_script()'
-script: R -q -e 'tic::script()'
-after_success: R -q -e 'tic::after_success()'
-after_failure: R -q -e 'tic::after_failure()'
-before_deploy: R -q -e 'tic::before_deploy()'
-after_script: R -q -e 'tic::after_script()'
-# DO NOT CHANGE THE CODE ABOVE
-
-language: r
+# OS ---------------------------------------------------------------------------
 os: osx
-cache:
-  - packages
-  - $HOME/.ccache
-  - $HOME/Library/Caches/Homebrew
-latex: false
-
-env:
-  global:
-  - MAKEFLAGS="-j 2"
-
