+++ man/TicStep.Rd
@@ -1,13 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/steps-base.R
-\docType{data}
 \name{TicStep}
 \alias{TicStep}
 \title{The base class for all steps}
-\format{An object of class \code{R6ClassGenerator} of length 24.}
-\usage{
-TicStep
-}
 \description{
 Override this class to create a new step.
 }
@@ -31,4 +26,45 @@ Return \code{FALSE} if conditions for running this step are not met.
 }
 }
 
-\keyword{datasets}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-new}{\code{TicStep$new()}}
+\item \href{#method-run}{\code{TicStep$run()}}
+\item \href{#method-prepare}{\code{TicStep$prepare()}}
+\item \href{#method-check}{\code{TicStep$check()}}
+}
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-new"></a>}}
+\subsection{Method \code{new()}}{
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{TicStep$new()}\if{html}{\out{</div>}}
+}
+
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-run"></a>}}
+\subsection{Method \code{run()}}{
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{TicStep$run()}\if{html}{\out{</div>}}
+}
+
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-prepare"></a>}}
+\subsection{Method \code{prepare()}}{
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{TicStep$prepare()}\if{html}{\out{</div>}}
+}
+
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-check"></a>}}
+\subsection{Method \code{check()}}{
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{TicStep$check()}\if{html}{\out{</div>}}
+}
+
+}
+}
