+++ inst/templates/circle-deploy-matrix.yml
@@ -1,4 +1,4 @@
-# tic documentation to get started: https://ropenscilabs.github.io/tic/
+# tic documentation: https://docs.ropensci.org/tic/dev/
 
 version: 2.1
 jobs:
@@ -12,11 +12,11 @@ jobs:
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
       - run: echo "$(date -d "$D" '+%d')r-release" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-checksum "/tmp/_tmp_file"
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
 
       # install deps and check pkg ---------------------------------------------
       - run:
-          name: Install package dependencies
+          name: Install dependencies
           command: |
             sudo apt update
             sudo apt install -y libglu1-mesa-dev libgmp-dev libgs-dev libgdal-dev libproj-dev libgeos-dev libudunits2-dev jags
@@ -27,7 +27,7 @@ jobs:
             R -q -e 'tic::before_install()'
             R -q -e 'tic::install()'
       - run:
-          name: Check package
+          name: R CMD Check
           no_output_timeout: 60m
           command: |
             R -q -e 'tic::before_script()'
@@ -35,7 +35,7 @@ jobs:
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
-          key: R-package-library-checksum "/tmp/_tmp_file"
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
           paths:
             - /usr/local/lib/R/site-library
   deploy:
@@ -47,14 +47,14 @@ jobs:
       - checkout
       - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-checksum "/tmp/_tmp_file"
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
       # Add your own SSH key fingerprint here, see https://circleci.com/docs/2.0/gh-bb-integration/#creating-a-github-deploy-key
       # and https://discuss.circleci.com/t/adding-ssh-keys-fails/7747/24 on how to create a valid key
       #- add_ssh_keys:
       #    fingerprints:
       #      - "15:9c:45:50:dc:94:87:6f:a9:ca:e1:fc:6a:b9:7e:62"
       - run:
-          name: Deploy Over SSH
+          name: Deploy
           command: |
             R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
             R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
@@ -73,11 +73,11 @@ jobs:
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
       - run: echo "$(date -d "$D" '+%d')3.5.3" > /tmp/_tmp_file
       - restore_cache:
-          key: R-package-library-checksum "/tmp/_tmp_file"
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
 
       # install deps and check pkg ---------------------------------------------
       - run:
-          name: Install package dependencies
+          name: Install dependencies
           command: |
             sudo apt update
             sudo apt install -y libglu1-mesa-dev libgmp-dev libgs-dev libgdal-dev libproj-dev libgeos-dev libudunits2-dev jags
@@ -88,7 +88,7 @@ jobs:
             R -q -e 'tic::before_install()'
             R -q -e 'tic::install()'
       - run:
-          name: Check package
+          name: R CMD Check
           no_output_timeout: 60m
           command: |
[…]
