+++ inst/templates/circle.yml
@@ -13,11 +13,11 @@ jobs:
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
       - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
       - restore_cache:
-         key: R-package-library-checksum "/tmp/_tmp_file"
+          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
 
       # install deps and check pkg ---------------------------------------------
       - run:
-          name: Install package dependencies
+          name: Install dependencies
           command: |
             echo -e 'options(Ncpus = 6, repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
             R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
@@ -26,7 +26,7 @@ jobs:
             R -q -e 'tic::before_install()'
             R -q -e 'tic::install()'
       - run:
-          name: Check package
+          name: R CMD Check
           no_output_timeout: 60m
           command: |
             R -q -e 'tic::before_script()'
@@ -34,7 +34,7 @@ jobs:
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
-          key: R-package-library-checksum "/tmp/_tmp_file"
+           key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
           paths:
             - /usr/local/lib/R/site-library
 
