+++ inst/templates/travis-deploy.yml
@@ -21,4 +21,3 @@ deploy:
 after_deploy: R -q -e 'tic::after_deploy()'
 after_script: R -q -e 'tic::after_script()'
 # DO NOT CHANGE THE CODE ABOVE -------------------------------------------------
-
