### 2019-11-30 341f035
use writeLines instead of use_tic_template() when writing single yml
