+++ R/use-yaml.R
@@ -21,79 +21,74 @@
 #'
 #'  | Provider | Operating system | Deployment | multiple R versions | Call |
 #'  | -------  | ---------------- | ---------- | ------------------- | ---- |
-#'  |----------|------------------|------------|---------------------|---------------------------------------------------------------|
-#'  | Travis   | Linux            | no         | no                  | `use_travis_yml("linux.yml")` |
-#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy.yml")` |
-#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix.yml")` |
-#'  |          | Linux            | yes        | yes                 | `use_travis_yml("linux-deploy-matrix.yml")` |
-#'  |          | macOS            | no         | no                  | `use_travis_yml("macOS.yml")` |
-#'  |          | macOS            | yes        | no                  | `use_travis_yml("macos-deploy.yml")` |
-#'  |          | macOS            | no         | yes                 | `use_travis_yml("macos-matrix.yml")` |
-#'  |          | macOS            | yes        | yes                 | `use_travis_yml("macos-deploy-matrix.yml")` |
-#'  |----------|------------------|------------|---------------------|---------------------------------------------------------------|
-#'  | Circle   | Linux            | no         | no                  | `use_circle_yml("linux.yml")` |
-#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy.yml")` |
-#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix.yml")` |'
-#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-deploy-matrix.yml")` |
-#'  |----------|------------------|------------|---------------------|---------------------------------------------------------------|
-#'  | Appveyor | Windows          | no         | no                  | `use_appveyor_yml("windows.yml")` |
-#'  |          | Windows          | no         | yes                 | `use_travis_yml("windows-matrix.yml")` |'
+#'  |----------|------------------|------------|---------------------|---------------------------------------------------------|
+#'  | Travis   | Linux            | no         | no                  | `use_travis_yml("linux")` |
+#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy")` |
+#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix")` |
+#'  |          | Linux            | yes        | yes                 | `use_travis_yml("linux-deploy-matrix")` |
+#'  |          | macOS            | no         | no                  | `use_travis_yml("macOS")` |
+#'  |          | macOS            | yes        | no                  | `use_travis_yml("macos-deploy")` |
+#'  |          | macOS            | no         | yes                 | `use_travis_yml("macos-matrix")` |
+#'  |          | macOS            | yes        | yes                 | `use_travis_yml("macos-deploy-matrix")` |
+#'  |----------|------------------|------------|---------------------|---------------------------------------------------------|
+#'  | Circle   | Linux            | no         | no                  | `use_circle_yml("linux")` |
+#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy")` |
+#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix")` |'
[…]
