### 2019-12-01 a1b3645
Ensure that _remotes_ is always the most recent version (#192)
* always upgrade remotes

* always install remotes fresh

* apply to templates
