### 2019-12-08 a5b309b
Update YAML templates and add table with options (#199)
* update templates

* use new yaml template approach

* update pkgdown reference

* update templates

* use writeLines instead of use_tic_template() when writing single yml

* add news

* always install remotes fresh
