+++ R/use-yaml.R
@@ -3,11 +3,11 @@
 #'
 #' @param type `[character]`\cr
 #'   Which template to use. The string should be given following the logic
-#'   `<provider>-<platform>-<action>`. See details for more.
+#'   `<platform>-<action>`. See details for more.
 #'
 #' @section Type:
 #' `tic` supports a variety of different YAML templates which follow the
-#'  `<provider>-<platform>-<action>` pattern. The first one is mandatory, the
+#'  `<platform>-<action>` pattern. The first one is mandatory, the
 #'  others are optional.
 #'
 #'  * Possible values for `<provider>` are `travis`, and `circle`
@@ -17,31 +17,78 @@
 #'  Not every combinations is supported on all CI systems.
 #'  For example, for `use_appveyor_yaml()` only `windows` and `windows-matrix` are valid.
 #'
+#'  Here is a list of all available combinations:
+#'
+#'  | Provider | Operating system | Deployment | multiple R versions | Call |
+#'  | -------  | ---------------- | ---------- | ------------------- | ---- |
+#'  |----------|------------------|------------|---------------------|---------------------------------------------------------|
+#'  | Travis   | Linux            | no         | no                  | `use_travis_yml("linux")` |
+#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy")` |
+#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix")` |
+#'  |          | Linux            | yes        | yes                 | `use_travis_yml("linux-deploy-matrix")` |
+#'  |          | macOS            | no         | no                  | `use_travis_yml("macOS")` |
+#'  |          | macOS            | yes        | no                  | `use_travis_yml("macos-deploy")` |
+#'  |          | macOS            | no         | yes                 | `use_travis_yml("macos-matrix")` |
+#'  |          | macOS            | yes        | yes                 | `use_travis_yml("macos-deploy-matrix")` |
+#'  |----------|------------------|------------|---------------------|---------------------------------------------------------|
+#'  | Circle   | Linux            | no         | no                  | `use_circle_yml("linux")` |
+#'  |          | Linux            | yes        | no                  | `use_travis_yml("linux-deploy")` |
+#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-matrix")` |'
+#'  |          | Linux            | no         | yes                 | `use_travis_yml("linux-deploy-matrix")` |
+#'  |----------|------------------|------------|---------------------|---------------------------------------------------------|
+#'  | Appveyor | Windows          | no         | no                  | `use_appveyor_yml("windows")` |
+#'  |          | Windows          | no         | yes                 | `use_travis_yml("windows-matrix")` |'
+#'
 #' @name yaml-templates
+#' @aliases yaml-templates
 #' @export
 use_travis_yml <- function(type) {
   if (type == "linux") {
-    template = "travis-linux.yml"
+    os = readLines(system.file("templates/travis-linux.yml", package = "tic"))
+    meta = readLines(system.file("templates/travis-meta-linux.yml", package = "tic"))
+    stages = readLines(system.file("templates/travis-no-deploy.yml", package = "tic"))
+    template = c(os, meta, stages)
   } else if (type == "linux-matrix") {
-    template = "travis-linux-matrix.yml"
+    os = readLines(system.file("templates/travis-linux.yml", package = "tic"))
[…]
