+++ man/prepare_all_stages.Rd
@@ -7,7 +7,7 @@
 prepare_all_stages(stages = dsl_load())
 }
 \arguments{
-\item{stages}{\code{[named list]}
+\item{stages}{\verb{[named list]}
 A named list of \code{TicStage} objects as returned by \code{\link[=dsl_load]{dsl_load()}},
 by default loaded from \code{tic.R}.}
 }
@@ -18,7 +18,8 @@ Run the \code{prepare()} method for all defined steps for which the
 \seealso{
 \link{TicStep}
 
-Other runners: \code{\link{run_all_stages}},
-  \code{\link{run_stage}}
+Other runners: 
+\code{\link{run_all_stages}()},
+\code{\link{run_stage}()}
 }
 \concept{runners}
