### 2019-12-08 #305 @pat-s
Hi @annakrystalli

_tic_ is now ready for migration to ropensci from our side.

The CRAN roadmap is as follows:

- Wait for _travis_ to go to ropensci and CRAN
- Wait for _cli_ v2.0 to go CRAN
- Wait for pat-s/circle to go to CRAN

Regarding _tic_ specifically:

- Implement a way to update to the latest version of the templates provided by _tic_
- Support Github Actions
- Support Azure pipelines
- memorize settings of `use_tic()` for use in multiple repos

We've done a major refactoring to how `use_tic()` works. It now guides the user through a set of questions and finally stores the selected yml templates in the repo as well as taking care for setting up permissions for deployment.
Also users can now single-call `use_XY_yml()` to simply add/overwrite their existing templates with the ones provided by _tic_ without having to go through the whole setup process of `use_tic()`.

We believe to have addressed all reviewer points and would like to thank again all people who have contributed to this. Sorry that everything took a bit long (from our side) and thanks for your understanding :)
