### 2019-12-08 fcb3e2e
apply tidyverse style, add .lintr config file
