+++ R/install.R
@@ -1,13 +1,13 @@
 # This code can only run as part of a CI run
 # nocov start
 
-verify_install <- function(pkg_names, pkgType = NULL) {
+verify_install <- function(pkg_names, pkgType = NULL) { # nolint
   # set "type" to platform default
-  pkgType <- update_type(pkgType)
+  pkgType <- update_type(pkgType) # nolint
   lapply(pkg_names, function(x) verify_install_one(x, pkgType = pkgType))
 }
 
-verify_install_one <- function(pkg_name, pkgType) {
+verify_install_one <- function(pkg_name, pkgType) { # nolint
   withr::with_options(
     c(pkgType = pkgType),
     remotes::install_cran(pkg_name, upgrade = TRUE)
