+++ R/steps-rcmdcheck.R
@@ -1,13 +1,13 @@
-RCMDcheck <- R6Class(
+RCMDcheck <- R6Class( # nolint
   "RCMDcheck",
   inherit = TicStep,
 
   public = list(
     initialize = function(warnings_are_errors = NULL, notes_are_errors = NULL,
-                              args = c("--no-manual", "--as-cran"),
-                              build_args = "--force", error_on = "warning",
-                              repos = repo_default(), timeout = Inf,
-                              check_dir = NULL) {
+                          args = c("--no-manual", "--as-cran"),
+                          build_args = "--force", error_on = "warning",
+                          repos = repo_default(), timeout = Inf,
+                          check_dir = NULL) {
       if (!is.null(notes_are_errors)) {
         warning_once(
           '`notes_are_errors` is deprecated, please use `error_on = "note"`'
@@ -135,7 +135,8 @@ RCMDcheck <- R6Class(
 step_rcmdcheck <- function(...,
                            warnings_are_errors = NULL, notes_are_errors = NULL,
                            args = NULL, build_args = NULL, error_on = "warning",
-                           repos = repo_default(), timeout = Inf, check_dir = NULL) {
+                           repos = repo_default(), timeout = Inf,
+                           check_dir = NULL) {
 
   #' @param build_args `[character]`\cr
   #'   Passed to `rcmdcheck::rcmdcheck()`.\cr
