+++ R/steps-base.R
@@ -55,10 +55,9 @@ HelloWorld <- R6Class(
 
 #' Step: Hello, world!
 #'
-#' The simplest step possible: prints "Hello, world!" to the console when run, does not require
-#' any preparation.
-#' This step may be useful to test a \pkg{tic} setup or as a starting point when implementing a
-#' custom step.
+#' The simplest step possible: prints "Hello, world!" to the console when run,
+#' does not require any preparation. This step may be useful to test a \pkg{tic}
+#' setup or as a starting point when implementing a custom step.
 #'
 #' @family steps
 #' @export
