+++ NEWS.md
@@ -1,3 +1,5 @@
+# tic 0.3.0
+
 - Same as previous version.
 
 # tic 0.2.13.9021
