+++ NEWS.md
@@ -1,3 +1,5 @@
+# tic (development version)
+
 # tic 0.3.0
 
 - add argument "check_dir" to step_rcmdcheck (#179)
