### 2019-12-10 #305 @annakrystalli
Hi @pat-s,

Thanks for the updates on the package! I'm excited to test out but I'm getting an error:
``` r
tic::use_tic()
#> Error in cli_alert("Welcome to {.pkg tic}!"): could not find function "cli_alert"
```

<sup>Created on 2019-12-10 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>.

Indeed running `devtools::check()` I get this note:
```
  use_tic: no visible global function definition for ‘cli_alert’
  use_tic: no visible global function definition for ‘cli_text’
  use_tic: no visible global function definition for ‘cli_h1’
  use_tic: no visible global function definition for ‘cli_ul’
  use_tic: no visible global function definition for ‘cli_par’
  use_tic: no visible global function definition for ‘cli_end’
  use_tic: no visible global function definition for ‘menu’
  yesno: no visible global function definition for ‘menu’
    cli_alert cli_end cli_h1 cli_par cli_text cli_ul menu
  Consider adding
    importFrom("utils", "menu")
  to your NAMESPACE file.
```

These functions are from package `cliapp`? At the minute, `cliapp` is not listed as a dependency, and the functions above don't seem to be imported anywhere. Could you have a little look into it?
