### 2019-12-10 6ffdd81
require cli 2.0.0
