### 2019-12-10 #305 @pat-s
@annakrystalli
Sorry for that. I missed to add the "cli (>= 2.0.0)" requirement in the DESCRIPTION. Before we only had "cli" there.

cli v2.0.0 is on CRAN since yesterday and installing tic should now get you the correct version during install.
