### 2019-12-10 d780061
remove use_travis_deploy(), lives in {travis} only
