+++ man/step_install_ssh_keys.Rd
@@ -15,8 +15,8 @@ Writes a private SSH key encoded in an environment variable
 to a file in \verb{~/.ssh}.
 Only run in non-interactive settings and if the environment variable
 exists and is non-empty.
-The \code{\link[=use_travis_deploy]{use_travis_deploy()}} and \code{\link[=use_tic]{use_tic()}} functions encode a private key
-as an environment variable for use with this function.
+The \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}} and \code{\link[=use_tic]{use_tic()}} functions encode a private
+key as an environment variable for use with this function.
 }
 \examples{
 dsl_init()
@@ -27,7 +27,7 @@ get_stage("before_deploy") \%>\%
 dsl_get()
 }
 \seealso{
-\code{\link[=use_travis_deploy]{use_travis_deploy()}}, \code{\link[=use_tic]{use_tic()}}
+\code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}, \code{\link[=use_tic]{use_tic()}}
 
 Other steps: 
 \code{\link{step_add_to_drat}()},
