+++ man/step_do_push_deploy.Rd
@@ -23,7 +23,7 @@ Default: deploy all files.}
 Commits and pushes to a repo prepared by \code{\link[=step_setup_push_deploy]{step_setup_push_deploy()}}.
 
 Deployment usually requires setting up SSH keys with
-\code{\link[=use_tic]{use_tic()}} or \code{\link[=use_travis_deploy]{use_travis_deploy()}}.
+\code{\link[=use_tic]{use_tic()}} or \code{\link[travis:use_travis_deploy]{travis::use_travis_deploy()}}.
 }
 \details{
 It is highly recommended to restrict the set of files
