+++ NAMESPACE
@@ -79,4 +79,5 @@ importFrom(crayon,has_color)
 importFrom(magrittr,"%>%")
 importFrom(memoise,memoise)
 importFrom(methods,setOldClass)
+importFrom(utils,menu)
 importFrom(utils,packageName)
