+++ inst/templates/travis-deploy.yml
@@ -1,13 +1,19 @@
-
+--- # DO NOT CHANGE THE CODE BELOW -------------------------------------------------
 # Stages -----------------------------------------------------------------------
 
-# DO NOT CHANGE THE CODE BELOW -------------------------------------------------
 before_install:
+  - echo -e "options(Ncpus = 8, repos = structure(c(CRAN = "https://cloud.r-project.org/")))" > $HOME/.Rprofile
+  - mkdir $HOME/.R && echo -e 'CXX_STD = CXX14\n\nCC=ccache gcc -std=gnu99\nCXX=ccache g++\nC11=ccache g++\nC14=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
+    mkdir $HOME/.ccache && echo -e 'max_size = 5.0G\nsloppiness = include_file_ctime\nhash_dir=false' > $HOME/.ccache/ccache.conf
+  - mkdir $HOME/.ccache && echo -e 'max_size = 5.0G\nsloppiness = include_file_ctime\nhash_dir=false' > $HOME/.ccache/ccache.conf
   - R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
   - R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
   - R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
   - R -q -e 'tic::before_install()'
-install: R -q -e 'tic::install()'
+install:
+  - if [ "${TRAVIS_OS_NAME}" == "osx" ]; then brew install ccache; fi
+  - if [ "${TRAVIS_OS_NAME}" == "osx" ]; then export PATH="/usr/local/opt/ccache/libexec:$PATH"; fi
+  - R -q -e 'tic::install()'
 before_script: R -q -e 'tic::before_script()'
 script: R -q -e 'tic::script()'
 after_success: R -q -e 'tic::after_success()'
