+++ inst/templates/travis-no-deploy.yml
@@ -1,8 +1,11 @@
-
+--- # DO NOT CHANGE THE CODE BELOW -------------------------------------------------
 # Stages -----------------------------------------------------------------------
 
-# DO NOT CHANGE THE CODE BELOW -------------------------------------------------
 before_install:
+  - echo -e "options(Ncpus = 8, repos = structure(c(CRAN = "https://cloud.r-project.org/")))" > $HOME/.Rprofile
+  - mkdir $HOME/.R && echo -e 'CXX_STD = CXX14\n\nCC=ccache gcc -std=gnu99\nCXX=ccache g++\nC11=ccache g++\nC14=ccache g++\nFC=ccache gfortran\nF77=ccache gfortran' > $HOME/.R/Makevars
+    mkdir $HOME/.ccache && echo -e 'max_size = 5.0G\nsloppiness = include_file_ctime\nhash_dir=false' > $HOME/.ccache/ccache.conf
+  - mkdir $HOME/.ccache && echo -e 'max_size = 5.0G\nsloppiness = include_file_ctime\nhash_dir=false' > $HOME/.ccache/ccache.conf
   - R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
   - R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
   - R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
