+++ inst/templates/travis-meta-linux.yml
@@ -1,4 +1,4 @@
---- # meta -------------------------------------------------------------------------
+--- # meta ---------------------------------------------------------------------
 language: r
 cache:
   - packages
