+++ inst/templates/circle-matrix.yml
@@ -99,3 +99,4 @@ workflows:
     jobs:
       - r-release
       - r-oldrelease
+# Custom user code -------------------------------------------------------------
