+++ inst/templates/circle-deploy.yml
@@ -85,3 +85,4 @@ workflows:
                 - master
     jobs:
       - r-release
+# Custom user code -------------------------------------------------------------
