+++ inst/templates/circle-deploy-matrix.yml
@@ -125,3 +125,4 @@ workflows:
     jobs:
       - r-release
       - r-oldrelease
+# Custom user code -------------------------------------------------------------
