### 2019-12-11 #305 @annakrystalli
Hey @pat-s,

Package installed successfully! 🎉
I'm trying to test it out with this example bookdown repo: https://github.com/annakrystalli/reprohack-docs

However, now at the end of `tic::use_tic()` using selections 2,1,2,2,4  I get:
```
Finished initiating sync with GitHub.
Waiting for sync with GitHub
Finished sync with GitHub.
Error: You may need to retry in a few seconds. If your repo annakrystalli/reprohack-docs belongs to an organization, you may need to allow Travis access to that organization. Please visit
  https://github.com/settings/connections/applications/f244293c729d5066cf27
  A browser window will be opened.
```

The page it launches though seems to show travis is sufficiently authorised?
![image](https://user-images.githubusercontent.com/5583057/70664609-a6db8b00-1c62-11ea-812d-a0de47a7e5f2.png)

Any idea why authorization is failing? It seemed to be working before.
