### 2019-12-11 #305 @pat-s
Hm, tricky. This check comes from `travis::travis_sync()`.

It is now easy to say what might be the problem here. You could

- call `travis::travis_sync()` manually again and it might resolve itself. Even though this should not be needed.
- try `ropenscilabs/travis@travis-com`. We did a complete refactoring of the auth process in this branch and it will soon be merged into master. This might also solve the issue already.

Other than that, I cannot reproduce the issue on my machine and we might dig deeper in a screenshare or similar if the issue persists.

However, the issue you encountered is in the end caused by the {travis} pkg and not by {tic}. Nevertheless I understand if you consider this as blocking because both packages are coupled closely, even though it is strictly triggered by an imported package.

Permission stuff with 3rd party providers is tricky as one cannot/hardly create automated tests for this. This is when user feedback becomes very important to sort things out, so thanks a lot!
