### 2019-12-15 #305 @annakrystalli
Hello again @pat-s 👋

OK, after installing the branch you suggested, things worked better. I still had a bit of trouble figuring out the setup. Finally, once I figured that the `R_TRAVIS` token is to specify the `endpoint` and I set it to `.com` it all worked.

The [build fails](https://travis-ci.com/annakrystalli/reprohack-docs/builds/141108558?utm_medium=notification&utm_source=email)

The pertinent error being:
```{bash}
$ R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
> {{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()
Error: object 'install_tic' not found
```

I reproduced this on another repo too

#### Additional comments

Like you say users need GITHUB authentication set up and travis (and any other service required) installed and authorised on GitHub (which can be even more complicated with orgs). As it is hard to automate troubleshooting all this, the docs will have to do the job.

While error messages are generally informative, some are more cryptic than others so the travisin the docs with a clear description of the various set up stages for each CI and anticipating searches for relevant information would be useful. Perhaps expanding on the `travis + tic` vignette and maybe even generalising to include the setup for other CIs too.

Ultimately, users shouldn't have to go to docs in other packages for basic set up instructions for using `tic`.

Here's some more specificing suggestions for improvements to documentation and functionality:

- [ ] Include description of correct travis app installation and permissions configuration on GitHub.
  - Include examples of properly configured pages (especially any launched during `use_tic`).
- [ ] mention the difference between `.com` and `.org`
  - (and relatedly the difference between the two github apps, `Travis CI` and `Travis CI for Open Source` which could also be confusing)
  - the [current ecosystem of change on travis](https://devops.stackexchange.com/questions/1201/whats-the-difference-between-travis-ci-org-and-travis-ci-com) with potentially links to migrating from `.org` to `.com`
- [ ] indeed, perhaps the default `R_TRAVIS` should be `.com` given the current changes to travis?
- [ ] what it is used for and how to set up `R_TRAVIS` token including an example of setting this up eg using `usthis::edit_r_environ()` to add env var to `.Renviron`
  - This could be cumbersome to work with if you currently have repositories split across both. Currently overriding it for individual repositories is possible through specifying the `endpoint` during `travis::travis_sync()` but this is not available through `use::tic()`. Perhaps it could be another question, an argument or some other way you think is most appropriate but I feel users should be able to override the default `endpoint` during `use::tic()`, at least for now.
- [ ] In a troubleshooting section, get in front of questions about:
  - common setup errors
  - [ ] setup for private repositories.
  - setup for orgs
- [ ] What do you think about including
  ```r
  use_tic
  ```
  without necessarily breaking down the questions and then a quick example of what a `tic.R` file might look like in the README? This allows code orriented users to get an idea of the workflow at first glance of the repo.

- [ ] During `travis::travis_sync()` the API token is pasted in the console and (although deleted form history) is still visible in the console. Could this be avoided in any way?

- [ ] When there is no github repository, I just get:

  ```
  Error: Failed to lookup git remotes
  ```
  Perhaps a more informative message, prompting the user to running (or even running) `usethis::use_github(protocol = "https")` (if using a PAT)

- [ ] I also noticed that I can run `dsl_init()` on repositories I have not initiated with `tic`. While running further functions (eg `do_bookdown()`) after that returns a non informative error. It would be better for `dsl_init()` to check whether the project is initialised with `tic.R` before trying to run macros.

- [ ] Docs url https://docs.ropensci.org/tic/dev resolves to a 404 page
