### 2019-12-15 #305 @pat-s
Hi @annakrystalli,

Thanks for the detailed review again. Important points, especially when it comes to user experience.

Do you consider **all** of these points (including some that only target the {travis} package) as blocking for the migration to _ropensci_? Maybe we can put them into issues of the respective repos to better address them?
Or are there just some of these (if at all) which you consider blocking for the migration?

(Migrating to _ropensci_ won't stop development therefore, especially because the package is not on CRAN yet. I am wondering if/how we can proceed with this submission here in the most effective way).
