### 2019-12-15 db5a67d
{{{install_tic}}} -> install_github()
