+++ inst/templates/appveyor-matrix.yml
@@ -13,7 +13,7 @@ install:
   - cmd: Rscript -e "getOption('repos')"
   - cmd: Rscript -e "if (!requireNamespace('remotes')) install.packages('remotes', type = 'binary')"
   - cmd: Rscript -e "if (getRversion() < '3.2' && !requireNamespace('curl')) install.packages('curl', type = 'binary')"
-  - cmd: Rscript -e "options(pkgType = 'binary'); {{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()"
+  - cmd: Rscript -e "options(pkgType = 'binary'); remotes::install_github('ropenscilabs/tic'); print(tic::dsl_load()); tic::prepare_all_stages()"
 
 before_build: Rscript -e "tic::before_install()"
 build_script: Rscript -e "tic::install()"
