### 2019-12-15 #305 @pat-s
> The build fails
>
> The pertinent error being:
>
> $ R -q -e '{{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()'
> > {{{install_tic}}}; print(tic::dsl_load()); tic::prepare_all_stages()
> Error: object 'install_tic' not found
> Execution halted
>

Thanks for catching this. Replacing `{{{install_tic}}}` with `remotes::install_github("ropenscilabs/tic")` solves this. I've also updated the templates already.
