### 2019-12-15 #305 @annakrystalli
OK thanks.

I've updated `.travis.yml` and pushed. Now I'm getting

```{bash}
> tic::install()
✔ Loading tic stage configuration from tic.R
Running install: step_install_deps(repos = repo_default())
<error/rlang_error>
cannot open the connection
Backtrace:
  1. tic::install()
  2. tic::run_stage("install", stages = stages)
  3. stage$run_all()
  4. private$run_one(step)
 11. step$run()
 14. remotes::install_deps(...)
 15. remotes::dev_package_deps(...)
 16. remotes:::load_pkg_description(pkgdir)
 17. remotes:::read_dcf(path_desc)
 20. base::read.dcf(path)
Error: A step failed in stage "install": install.
In addition: Warning message:
In read.dcf(path) :
  cannot open compressed file '/home/travis/build/annakrystalli/reprohack-docs/DESCRIPTION', probable reason 'No such file or directory'
Execution halted
The command "R -q -e 'tic::install()'" failed and exited with 1 during .
```

I'm just using tic vanilla bookdown setup https://github.com/annakrystalli/reprohack-docs. Is a DESCRIPTION file necessary for this to work?
