### 2019-12-15 #305 @pat-s
> Is a DESCRIPTION file necessary for this to work?

Yes, because {{tic}} installs the dependencies based on that. No matter the project type, there always needs to be a DESCRIPTION file.
See also the [bookdown example repo](https://github.com/ropenscilabs/tic.bookdown) for more info. I think the default error message is quite descriptive here, what do you think?
