### 2019-12-15 #305 @annakrystalli
SUCCESS!! 🎉

So, If the project requires it, I feel `do_bookdown()` should check for a `DESCRIPTION` file and warn earlier rather than wait to get an error on TRAVIS. Indeed the vanilla new bookdown project (created through Rstudio) which is how I set up the test repo, does not come with a `DESCRIPTION` file.

If I try and use `usethis::use_description()` to get a starting template, I get an error because the repo name `reprohack-docs` has non ascii character (which might be typical for non package projects). I ended up using `holepunch::write_compendium_description()` to create a `DESCRIPTION` file but it might be nice to have similar functionality within `tic` and create a `DESCRIPTION` file if it's missing during `use_tic`, given that it's effectively a dependency for any `tic` project.

Indeed, if `tic` is guessing which type of project it is, it could also check to make sure known required dependencies are specified (eg markdown, bookdown etc).  it should also be mentioned in function/general docs that a `DESCRIPTION` file is required and should be used to track dependencies. Might be good to mention that missing dependencies in the `DESCRIPTION` file will result in errors on travis.

Regarding your previous questions, I don't feel review can be complete without some of the documentation issues being addressed. It took a lot of time to try and get what should have been a simple test project so I imagine that many people might find setup confusing as it stands.

I also can't accept `tic` until the version of `travis` I had to install is merged into master. Do you have any idea when that might be?
