### 2019-12-15 ropenscilabs/tic#200 @pat-s
`use_tic()`: Ensure that a DESCRIPTION file is present
Otherwise, macro functions will error during the CI run. See https://github.com/ropensci/software-review/issues/305#issuecomment-565823673.

- [ ] Assert a DESCRIPTION file in `use_tic()`. Create one if missing? Repo names with `-` are problematic.

- [ ] Mention (more clearly) that build dependencies are installed from packages listed in the DESCRIPTION file
