### 2019-12-15 #305 @pat-s
> So, If the project requires it, I feel do_bookdown() should check for a DESCRIPTION file and warn earlier rather than wait to get an error on TRAVIS. Indeed the vanilla new bookdown project (created through Rstudio) which is how I set up the test repo, does not come with a DESCRIPTION file.

I see. We can check for this during `use_tic()` but not within `do_bookdown()`. The latter is a macro function and is only being executed on the CI system (unless the user emulates the CI run locally). However, it does not have any setup/check functionality before starting a run.

I've created an issue to better assert the existence of a DESCRIPTION file.

> I also can't accept tic until the version of travis I had to install is merged into master. Do you have any idea when that might be?

The refactoring should be done within the next days/weeks. In general, {tic} should also work with the current state of the {travis} package, even though it is limited to the .org endpoint only and its usage is a bit more annoying than the refactored version in the `travis-com` branch.

Could we quickly summarize which issues are blocking for the transition? I'd like to outsource your points from https://github.com/ropensci/software-review/issues/305#issuecomment-565804724 into issues in the repo and label them accordingly. You might as well do it on your own if you want to add more context to certain points :)
