### 2019-12-15 #305 @annakrystalli
I think you should incorporate all the comments as issues as you feel they make sense.  I've ticked the ones I would consider necessary to pass review, the rest you can add as enhancements. But it would be good to get a full response to the points raised once you have made changes.

> In general, {tic} should also work with the current state of the {travis} package, even though it is limited to the .org endpoint

The problem with that is that since May 2018, travis has [announced a movement towards `.com`]( https://blog.travis-ci.com/2018-05-02-open-source-projects-on-travis-ci-com-with-github-apps) so I'd rather not accept `tic` without `.com` endpoint functionality.
