### 2019-12-22 8c90797
fix travis templates, fixes #201
