+++ inst/templates/travis-meta-linux.yml
@@ -1,4 +1,5 @@
---- # meta ---------------------------------------------------------------------
+
+# meta ---------------------------------------------------------------------
 language: r
 cache:
   - packages
