+++ inst/templates/travis-deploy.yml
@@ -1,4 +1,4 @@
---- # DO NOT CHANGE THE CODE BELOW ---------------------------------------------
+
 # Stages -----------------------------------------------------------------------
 
 before_install:
@@ -25,6 +25,5 @@ deploy:
     all_branches: true
 after_deploy: R -q -e 'tic::after_deploy()'
 after_script: R -q -e 'tic::after_script()'
-# DO NOT CHANGE THE CODE ABOVE -------------------------------------------------
 
 # Custom user code -------------------------------------------------------------
