### 2020-01-01 5b7690a
Add arguments to use_tic()
