+++ R/use_tic.R
@@ -11,9 +11,38 @@
 #' 1. Create a default `tic.R` file depending on the repo type
 #'    (package, website, bookdown, ...)
 #'
+#' @param wizard `[flag]`\cr Interactive operation? If `TRUE`,
+#'   a menu will be shown, and the `linux`, `mac`, `windows`,
+#'   `deploy` and `matrix` arguments are ignored.
+#' @param linux,windows,mac `[string]`\cr Which CI provider(s) to use to test on
+#'   Linux, Windows, and macOS, respectively.
+#' @param deploy `[string]`\cr Which CI provider(s) to use to
+#'   deploy artifacts such as pkgdown documentation.
+#' @param matrix `[string]`\cr For which CI provider(s) to set up
+#'   matrix builds.
 #' @param quiet `[flag]`\cr Less verbose output? Default: `FALSE`.
 #' @export
-use_tic <- function(quiet = FALSE) { # nolint
+#' @examples
+#' # Requires interactive mode
+#' if (FALSE) {
+#'   use_tic()
+#'
+#'   # Pre-specified settings favoring Circle CI:
+#'   use_tic(
+#'     linux = "circle",
+#'     mac = "travis",
+#'     windows = "appveyor",
+#'     deploy = "circle",
+#'     matrix = "all"
+#'   )
+#' }
+use_tic <- function(wizard = interactive(),
+                    linux = c("travis", "circle", "none", "all"),
+                    mac = c("none", "travis", "all"),
+                    windows = c("none", "appveyor", "all"),
+                    deploy = c("travis", "circle", "none", "all"),
+                    matrix = c("none", "travis", "circle", "appveyor", "all"),
+                    quiet = FALSE) { # nolint
   cli_alert("Welcome to {.pkg tic}!")
   cli_text(
     "This wizard will guide you through the setup process for getting started
@@ -35,37 +64,62 @@ use_tic <- function(quiet = FALSE) { # nolint
   cli_end()
 
   cli_text("We recommend the following setup:")
-  cli_ul(c("Appveyor: Windows", "Circle CI: Linux", "Travis CI: macOS"))
+  cli_ul(c("Travis CI: Linux", "Travis CI: macOS", "Appveyor: Windows"))
   cli_par()
   cli_end()
 
-  if (yesno("Ready to get started?")) {
-    return(invisible(NULL))
+  if (wizard) {
+    if (yesno("Ready to get started?")) {
+      return(invisible(NULL))
+    }
+
+    cli_h1("Choosing your setup.")
+    cli_text("We'll ask you a few yes/no questions to gather your preferences.")
+    cli_par()
+    cli_end()
+
+    linux <- ci_menu(linux,
+      title = "Which provider do you want to use for Linux builds?"
+    )
+
+    mac <- ci_menu(mac,
+      title = "Do you want to build on macOS (= Travis CI)?"
+    )
+
+    windows <- ci_menu(windows,
+      title = "Do you want to build on Windows (= Appveyor)?"
+    )
+
+    deploy <- ci_menu(intersect(deploy, c(linux, mac, windows, "all", "none")),
+      title = "Do you want to deploy (i.e. push from the CI build to your repo) on certain providers? If yes, which ones?" # nolint
+    )
+
+    matrix <- ci_menu(intersect(matrix, c(linux, mac, windows, "all", "none")),
+      title = "Do you want to build on multiple R versions? (i.e. R-devel, R-release, R-oldrelease). If yes, on which platform(s)?" # nolint
+    )
+
+    wizard <- FALSE
+    use_tic_call <- c(
+      'tic::use_tic(',
[…]
