### 2020-01-01 6a9a7f7
Remove all if only one true choice
