### 2020-01-01 6ba9ab7
Compact use_tic() call
