### 2020-01-02 31aaa68
precommit and lintr
