### 2020-01-02 3d45e4b
add option to combine macos and linux on travis
