### 2020-01-05 0fc6886
fix use_appveyor_yml()
