### 2020-01-05 5bc354b
update badges of repo examples
