### 2020-01-06 503e1f1
fix broken link + more accessible link
https://webaccess.berkeley.edu/ask-pecan/click-here
