### 2020-01-06 db5b706
add_code_step: skip pkg installation autodetection if prepare_call is given
