+++ R/steps-code.R
@@ -15,15 +15,15 @@ RunCode <- R6Class(
     },
 
     prepare = function() {
-      # Needs to happen before auto-detection of package to be installed,
-      # to allow installation of packages from nonstandard repositories
+      # allow installation of packages from nonstandard repositories, e.g.
+      # Github packages using a repo slug
       if (!is.null(private$prepare_call)) {
         private$install_call_dep(private$prepare_call)
         set.seed(private$seed)
         eval(private$prepare_call, envir = .GlobalEnv)
+      } else {
+        private$install_call_dep(private$call)
       }
-
-      private$install_call_dep(private$call)
     }
   ),
 
