### 2020-01-06 c83415f
update doc for TicStep
