### 2020-01-06 3360749
fix broken sentence in README, fixes #209
