### 2020-01-07 c94475c
make it possible to run `use_tic(wizard = FALSE) with proper defaults
