### 2020-01-07 59a3818
add new line to circle templates
