+++ inst/templates/circle-deploy.yml
@@ -86,3 +86,4 @@ workflows:
     jobs:
       - r-release
 # Custom user code -------------------------------------------------------------
+
