### 2020-01-07 3523d78
add ccache section for macOS on Travis
