### 2020-01-07 83d9bd4
- use_tic(): Travis as default for Linux and macOS
