+++ R/use_tic.R
@@ -44,10 +44,10 @@
 #'   )
 #' }
 use_tic <- function(wizard = interactive(),
-                    linux = "circle",
-                    mac = "none",
+                    linux = "travis",
+                    mac = "travis",
                     windows = "appveyor",
-                    deploy = "circle",
+                    deploy = "travis",
                     matrix = "none",
                     quiet = FALSE) { # nolint
 
