### 2020-01-07 814d31d
check for valid DESCRIPTION
