### 2020-01-07 3e50f06
make it possible to pass the endpoint arg to travis funs
