### 2020-01-07 e58b719
only run steps and macros interactively
