+++ R/steps-base.R
@@ -73,5 +73,8 @@ HelloWorld <- R6Class(
 #'
 #' dsl_get()
 step_hello_world <- function() {
+  if (interactive()) {
+    stop("step_* functions should only be used in tic.R and not interactively.")
+  }
   HelloWorld$new()
 }
