+++ R/steps-bookdown.R
@@ -38,5 +38,8 @@ BuildBookdown <- R6Class(
 #'
 #' dsl_get()
 step_build_bookdown <- function(...) {
+  if (interactive()) {
+    stop("step_* functions should only be used in tic.R and not interactively.")
+  }
   BuildBookdown$new(...)
 }
