+++ R/macro-bookdown.R
@@ -37,6 +37,10 @@ do_bookdown <- function(...,
                         remote_url = NULL,
                         commit_message = NULL, commit_paths = ".") {
 
+  if (interactive()) {
+    stop("Macro functions should only be used in tic.R and not interactively.")
+  }
+
   #' @param deploy `[flag]`\cr
   #'   If `TRUE`, deployment setup is performed
   #'   before building the bookdown site,
