### 2020-01-07 f7e8a40
ensure that tic.R exists when calling dsl_init()
