### 2020-01-07 4a6b195
fix checking for DESCRIPTION
