+++ R/stage.R
@@ -36,7 +36,7 @@ TicStage <- R6Class( # nolint
 
       # check if DESCRIPTION file exists
       path <- paste0(usethis::proj_get(), "/DESCRIPTION")
-      if (!file_exists(path)) {
+      if (!file.exists(path)) {
         cli::cli_alert_danger("{.pkg tic} requires a DESCRIPTION file for
         installing dependencies, please add one.", wrap = TRUE)
         stop("Can't find DESCRIPTION.",
