### 2020-01-07 cd333e6
fix stage.R
