+++ R/utils.R
@@ -64,7 +64,7 @@ valid_package_name <- function(x) {
   grepl("^[a-zA-Z][a-zA-Z0-9.]+$", x) && !grepl("\\.$", x)
 }
 
-project_name <- function(base_path = proj_get()) {
+project_name <- function(base_path = usethis::proj_get()) {
   if (is_package(base_path)) {
     project_data(base_path)$Package
   } else {
@@ -72,7 +72,7 @@ project_name <- function(base_path = proj_get()) {
   }
 }
 
-project_data <- function(base_path = proj_get()) {
+project_data <- function(base_path = usethis::proj_get()) {
   if (is_package(base_path)) {
     data <- package_data(base_path)
   } else {
@@ -81,7 +81,7 @@ project_data <- function(base_path = proj_get()) {
   data
 }
 
-is_package <- function(base_path = proj_get()) {
+is_package <- function(base_path = usethis::proj_get()) {
   res <- tryCatch(
     rprojroot::find_package_root_file(path = base_path),
     error = function(e) NULL
@@ -89,7 +89,7 @@ is_package <- function(base_path = proj_get()) {
   !is.null(res)
 }
 
-package_data <- function(base_path = proj_get()) {
+package_data <- function(base_path = usethis::proj_get()) {
   desc <- desc::description$new(base_path)
   as.list(desc$get(desc$fields()))
 }
