### 2020-01-08 cd722cb
try running tic functions within repo dir
