### 2020-01-08 f2f9980
run tests in repo dir
