+++ DESCRIPTION
@@ -55,14 +55,17 @@ Suggests:
     callr,
     circle,
     covr,
+    desc,
     devtools,
     drat,
     fansi,
     git2r,
     knitr,
     openssl,
+    pkgdown,
     rcmdcheck,
     rmarkdown,
+    rprojroot,
     testthat (>= 2.1.0),
     travis (>= 0.2.11.9001),
     usethis
