+++ man/dsl.Rd
@@ -50,9 +50,11 @@ and the links therein for available steps.
 \code{add_code_step()} is a shortcut for \code{add_step(step_run_code(...))}.
 }
 \examples{
+\dontrun{
 dsl_init()
 
 get_stage("script")
+}
 
 get_stage("script") \%>\%
   add_step(step_hello_world())
