+++ R/dsl.R
@@ -48,11 +48,12 @@ get_stage <- function(name) {
 #' @rdname dsl
 #' @export
 #' @examples
-#'
+#' \dontrun{
 #' get_stage("script") %>%
 #'   add_step(step_hello_world())
 #'
 #' get_stage("script")
+#' }
 add_step <- function(stage, step) {
   step_quo <- enquo(step)
 
@@ -80,11 +81,12 @@ add_step <- function(stage, step) {
 #' @inheritParams step_run_code
 #' @rdname dsl
 #' @examples
-#'
+#' \dontrun{
 #' get_stage("script") %>%
 #'   add_code_step(print("Hi!"))
 #'
 #' get_stage("script")
+#' }
 add_code_step <- function(stage, call = NULL, prepare_call = NULL) {
   call_expr <- enexpr(call)
   prepare_call_expr <- enexpr(prepare_call)
