+++ R/steps-write-text-file.R
@@ -36,12 +36,14 @@ WriteTextFile <- R6Class(
 #' @family steps
 #' @export
 #' @examples
+#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script") %>%
 #'   add_step(step_write_text_file("Hi!", path = "hello.txt"))
 #'
 #' dsl_get()
+#' }
 step_write_text_file <- function(..., path) {
   if (interactive()) {
     stop("step_* functions should only be used in tic.R and not interactively.")
