+++ R/steps-rcmdcheck.R
@@ -126,12 +126,14 @@ RCMDcheck <- R6Class( # nolint
 #'   check. Defaults to project root for easy upload of artifacts.
 #' @export
 #' @examples
+#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script") %>%
 #'   add_step(step_rcmdcheck(error_on = "note", repos = repo_bioc()))
 #'
 #' dsl_get()
+#' }
 step_rcmdcheck <- function(...,
                            warnings_are_errors = NULL, notes_are_errors = NULL,
                            args = NULL, build_args = NULL, error_on = "warning",
