+++ R/steps-drat.R
@@ -29,12 +29,14 @@ AddToDrat <- R6Class(
 #' @family steps
 #' @export
 #' @examples
+#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script") %>%
 #'   add_step(step_add_to_drat())
 #'
 #' dsl_get()
+#' }
 step_add_to_drat <- function() {
   if (interactive()) {
     stop("step_* functions should only be used in tic.R and not interactively.")
