+++ R/steps-base.R
@@ -66,12 +66,14 @@ HelloWorld <- R6Class(
 #' @family steps
 #' @export
 #' @examples
+#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script") %>%
 #'   add_step(step_hello_world())
 #'
 #' dsl_get()
+#' }
 step_hello_world <- function() {
   if (interactive()) {
     stop("step_* functions should only be used in tic.R and not interactively.")
