+++ DESCRIPTION
@@ -5,7 +5,7 @@ Title: Tasks Integrating Continuously: CI-Agnostic Workflow
 Version: 0.3.0.9002
 Authors@R: 
     c(person(given = "Kirill",
-             family = "M\u00fcller",
+             family = "Müller",
              role = c("aut", "cre"),
              email = "krlmlr+r@mailbox.org",
              comment = c(ORCID = "0000-0002-1416-3412")),
@@ -48,6 +48,7 @@ Imports:
     R6,
     remotes,
     rlang (>= 0.4.1.9000),
+    usethis,
     withr
 Suggests: 
     base64enc,
@@ -67,8 +68,7 @@ Suggests:
     rmarkdown,
     rprojroot,
     testthat (>= 2.1.0),
-    travis (>= 0.2.11.9001),
-    usethis
+    travis (>= 0.2.11.9001)
 VignetteBuilder:
     knitr
 Remotes: 
