+++ R/stage.R
@@ -37,7 +37,7 @@ TicStage <- R6Class( # nolint
       # check if DESCRIPTION file exists
       path <- paste0(usethis::proj_get(), "/DESCRIPTION")
       if (!file.exists(path)) {
-        cli::cli_alert_danger("{.pkg tic} requires a DESCRIPTION file for
+        cli_alert_danger("{.pkg tic} requires a DESCRIPTION file for
         installing dependencies, please add one.", wrap = TRUE)
         stop("Can't find DESCRIPTION.",
           call. = FALSE
