### 2020-01-08 #305 @pat-s
> Include description of correct travis app installation and permissions configuration on GitHub.

There is only one app left that can be actively installed. I written some sentences about this in https://github.com/ropenscilabs/tic/pull/213

> mention the difference between .com and .org

Done in https://github.com/ropenscilabs/tic/pull/213

> indeed, perhaps the default R_TRAVIS should be .com given the current changes to travis?

Done in https://github.com/ropenscilabs/travis/commit/a8b7ad37ad1a56fcdc34f03d542bead248decd18

> what it is used for and how to set up R_TRAVIS token including an example of setting this up eg using usthis::edit_r_environ() to add env var to .Renviron

I don't think that this belong into the {tic} package but rather into the {travis} package. The {tic} package has already so many vignettes and we need to take care not to bloat up the package too much. We can't write a vignette about the specialties of all existing and upcoming backends. There is already a section about how to set `R_TRAVIS` at https://docs.ropensci.org/travis/articles/travis.html#setting-the-default-endpoint.

> This could be cumbersome to work with if you currently have repositories split across both. Currently overriding it for individual repositories is possible through specifying the endpoint during travis::travis_sync() but this is not available through use::tic(). Perhaps it could be another question, an argument or some other way you think is most appropriate but I feel users should be able to override the default endpoint during use::tic(), at least for now.

I see your concern here but disagree. We would like not to make an opinionated statement which endpoint to use and if one should fully migrate or not. This is a decision the user needs to make.

Regarding passing the `endpoint` arg in `use_tic()`: https://github.com/ropenscilabs/tic/pull/213

> In a troubleshooting section, get in front of questions about:
>
>     - common setup errors
>     - setup for private repositories.
>     - setup for orgs

I am not sure about such a vignette. Most people prefer searching error messages and not many people will read a vignette upfront before trying out the functions. Also I would not know right now what to put there.
Regarding the other two points: There is no difference for these scenarios. If so, issues should go to the API client packages {travis} and {circle}.

> without necessarily breaking down the questions and then a quick example of what a tic.R file might look like in the README? This allows code orriented users to get an idea of the workflow at first glance of the repo.

There is a small example in https://docs.ropensci.org/tic/articles/tic.html already. I am not sure what you mean by "not breaking down the questions"? We hope that `?tic::use_tic()` helps. I am not sure to what extent a c/p tic.R in the README would add added value compared to all the contents in the already existing vignettes. The vignettes break down how {tic} works, what steps/stages are and that everything comes together in the `tic.R` file. Is there something you miss in all of these? Or could you maybe clarify more what the advantage of this action would be so I can follow?

> When there is no github repository, I just get:

I think users interested in a package simplifying CI usage are familiar enough with git/Github to know that their local repo needs to be linked. The error message is even quite descriptive imo. I don't think that checking this is within the responsibility of this package.

> I also noticed that I can run dsl_init() on repositories I have not initiated with tic. While running further functions (eg do_bookdown()) after that returns a non informative error. It would be better for dsl_init() to check whether the project is initialised with tic.R before trying to run macros.

{tic} macro-functions and steps (i.e. all funs starting with "do*" or "step" are not aimed to be run standalone. We now return early if those functions are used interactively. Also we now suggest to call `use_tic()` first if someone calls `dsl_init()` without a `tic.R` file (https://github.com/ropenscilabs/tic/pull/213)

> I also can't accept tic until the version of travis I had to install is merged into master. Do you have any idea when that might be?

The {travis} refactoring is done.

> Docs url docs.ropensci.org/tic/dev resolves to a 404 page

We will stick with https://docs.ropensci.org/tic until we are on CRAN.
