### 2020-01-10 2dc3a45
circle templates: Call "apt update" before ccache install, use `mkdir -p` instead of `mkdir`
