### 2020-01-11 d6ffd8f
* ci_can_push(): Change default from id_rsa to TRAVIS_DEPLOY_KEY but also support backward comp
