+++ R/appveyor.R
@@ -29,7 +29,13 @@ AppVeyorCI <- R6Class( # nolint
     get_commit = function() {
       Sys.getenv("TRAVIS_COMMIT")
     },
-    can_push = function(name = "id_rsa") {
+    can_push = function(name = "TRAVIS_DEPLOY_KEY") {
+      # id_rsa is the "old" name which was previously hard coded in the {travis}
+      # package. New default name: "TRAVIS_DEPLOY_KEY"
+      # for backward comp we check for the old one too
+      if (self$has_env("id_rsa")) {
+        return(invisible())
+      }
       self$has_env(name)
     },
     get_env = function(env) {
