+++ R/mock.R
@@ -25,7 +25,7 @@ MockCI <- R6Class(
     get_commit = function() {
       "00000000000000000000000000000000"
     },
-    can_push = function(name = "id_rsa") {
+    can_push = function(name = "TRAVIS_DEPLOY_KEY") {
       self$has_env(name)
     },
     get_env = function(env) {
