### 2020-01-11 94ecc33
Typo fix for travis matrix: oldlrease -> oldrel
On Travis the old R release is referred to as `oldrel`. Using `oldrelease` results in the following completely uninformative build error:

```
There was an error in the .travis.yml file from which we could not recover.

Unfortunately, we do not know much about this error.

Please review https://docs.travis-ci.com, or contact us at support@travis-ci.com with the error ID: XXX
```
