### 2020-01-11 42addcf
* optimize travis templates by accounting which job is used for the pkgdown build
