+++ inst/templates/travis-meta-linux.yml
@@ -5,7 +5,3 @@ cache:
   - packages
   - ccache
 latex: false
-
-env:
-  global:
-    - MAKEFLAGS="-j 2"
