+++ inst/templates/travis-env-no-pkgdown.yml
@@ -1,0 +1,6 @@
+
+# environment variables --------------------------------------------------------
+
+env:
+  global:
+    - MAKEFLAGS="-j 2"
