### 2020-01-11 a46d05c
* fix ci_can_push() for travis
