### 2020-01-11 1b150cd
try fixing step_test_ssh()
