### 2020-01-12 fc68829
pass down name arg to initialize function of TestSSH
