### 2020-01-12 392dc1b
do the renaming in install_ssh_key and not in test_ssh_key
