+++ man/step_test_ssh.Rd
@@ -4,11 +4,7 @@
 \alias{step_test_ssh}
 \title{Step: Test SSH connection}
 \usage{
-step_test_ssh(
-  url = "git@github.com",
-  verbose = "-v",
-  name = "TRAVIS_DEPLOY_KEY"
-)
+step_test_ssh(url = "git@github.com", verbose = "-v")
 }
 \arguments{
 \item{url}{\verb{[string]}\cr
@@ -16,10 +12,6 @@ URL to establish SSH connection with, by default \code{git@github.com}}
 
 \item{verbose}{\verb{[string]}\cr
 Verbosity, by default \code{"-v"}. Use \code{"-vvv"} for more verbosity.}
-
-\item{name}{\verb{[string]}\cr
-Name of the environment variable and the target file, default:
-\code{"TRAVIS_DEPLOY_KEY"}.}
 }
 \description{
 Establishes an SSH connection.
