### 2020-01-12 5030d2f
fix permissions of ssh key
