### 2020-01-12 e3f47bb
try to tell git and others which ssh key to use
