### 2020-01-12 2d490c0
fix path to ssh key pair when testing connection
