### 2020-01-12 3410f63
add ssh key to agent and be more verbose
