### 2020-01-12 1bed082
start ssh agent before adding the key
