+++ R/steps-ssh.R
@@ -100,6 +100,7 @@ InstallSSHKeys <- R6Class(
       )
 
       # add ssh key to agent
+      system2("eval", c("`ssh-agent -s`"))
       message("Adding ssh key to ssh-agent.")
       system2("ssh-add", c(file.path("~", ".ssh", name)))
     },
