### 2020-01-12 e4f4e81
next try to start the ssh agent
