### 2020-01-12 a1c26eb
try adding support for old 'id_rsa' name
