+++ R/steps-ssh.R
@@ -105,7 +105,7 @@ InstallSSHKeys <- R6Class(
       )
 
       # add ssh key to agent
-      system2("ssh-agent", "/bin/sh")
+      system("eval `ssh-agent -s`")
       message("Adding ssh key to ssh-agent.")
       system2("ssh-add", c(file.path("~", ".ssh", name)))
     },
