+++ R/steps-ssh.R
@@ -105,8 +105,8 @@ InstallSSHKeys <- R6Class(
       )
 
       # add ssh key to agent
-      system("eval `ssh-agent -s`")
-      message("Adding ssh key to ssh-agent.")
+      # system("eval `ssh-agent -s`")
+      # message("Adding ssh key to ssh-agent.")
       system2("ssh-add", c(file.path("~", ".ssh", name)))
     },
 
