### 2020-01-12 aa32762
use {pkg} notation for pkgs
