### 2020-01-12 9dc1bc2
also pass travis_private_key_name to ci_can_push()
