### 2020-01-12 f6fa788
correctly pass the private_key_name default
