### 2020-01-12 9acc507
vignettes: better document latest deployment changes, use {pkg} notation for pkgs
