+++ vignettes/deployment.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 
 ## Introduction
 
-_tic_ uses the [CI client packages](ci-client-packages.html) {travis} and {circle} for setting up deployment on the CI systems.
+{tic} uses the [CI client packages](ci-client-packages.html) {travis} and {circle} for setting up deployment on the CI systems.
 
 ### Travis CI
 
@@ -27,7 +27,8 @@ _tic_ uses the [CI client packages](ci-client-packages.html) {travis} and {circl
     knitr::include_graphics("img/ssh-deploy-key.png")
     ```
     
-2. A private key is encoded in base64 and stored in an environment variable on Travis. During the CI run, _tic_ installs it in the `~/.ssh` directory so that it can be used for authentication:
+2. A private key is encoded in base64 and stored in an environment variable on Travis. 
+   During the CI run, {tic} installs it in the `~/.ssh` directory so that it can be used for authentication:
 
     ```{r, echo = FALSE, fig.align='center', dpi = 10}
     knitr::include_graphics("img/private-key-travis.png")
@@ -39,7 +40,7 @@ _tic_ uses the [CI client packages](ci-client-packages.html) {travis} and {circl
 
 On Circle CI, setting up deployment is easier than on Travis as there is no need to create a SSH key pair for deployment.
 
-When calling `circle::use{circle}deploy()` directly (or indirectly via `tic::use_tic()`), a so called "github-user-key" is created and stored in the Circle CI repo.
+When calling `circle::use_circle_deploy()` directly (or indirectly via `tic::use_tic()`), a so called "user-key" is created and stored in the Circle CI repo.
 This key makes it possible to deploy from Circle CI builds to your Github repo.
 No extra deploy key need to be stored in the Github repo.
 
@@ -48,12 +49,12 @@ No extra deploy key need to be stored in the Github repo.
 [`pkgdown`](https://github.com/r-lib/pkgdown) is an R package which builds a documentation wrapper-site of a R package.
 It collects all the vignettes, function references and metadata information of a package and presents it in an eye-appealing HTML version.
 
-It has become a quasi-standard to provide a _pkgdown_ site for an R package.
+It has become a quasi-standard to provide a {pkgdown} site for an R package.
 However, it is tedious to update the pkgdown site manually on every commit, check whether something has changed and commit the changes.
-_tic_ comes with the ability to automate this procedure.
+{tic} comes with the ability to automate this procedure.
 
-The following example shows how _tic_ deploys a `pkgdown` site on Travis CI.
-In the `.travis.yml` file the "before_deploy" and "deploy" stages are redirected to _tic_.
+The following example shows how {tic} deploys a `pkgdown` site on Travis CI.
+In the `.travis.yml` file the "before_deploy" and "deploy" stages are redirected to {tic}.
 
 ```{yml eval = FALSE}
 before_deploy: R -q -e 'tic::before_deploy()'
@@ -62,17 +63,24 @@ deploy:
   script: R -q -e 'tic::deploy()'
 ```
 
-In the "before_deploy" stage, _tic_ will do the following:
+In the "before_deploy" stage, {tic} will do the following:
 
 ```{r eval = FALSE}
-if (ci_has_env("BUILD_PKGDOWN")) {
+if (ci_on_travis()) {
   get_stage("before_deploy") %>%
-    add_step(step_setup_ssh())
+    add_step(step_setup_ssh(name = travis_private_key_name))
 }
 ```
 
-- Call `step_setup_ssh()` if the environment variable `"BUILD_PKGDOWN"` is set in the Travis build. This step sets up SSH key previously added to the Github and Travis via `travis::use_travis_deploy()`.
[…]
