+++ vignettes/advanced.Rmd
@@ -18,14 +18,14 @@ knitr::opts_chunk$set(
 
 ## Running steps conditionally
 
-Steps and stages can be run conditionally using the control workflow of _tic_.
+Steps and stages can be run conditionally using the control workflow of {tic}.
 Possible conditionals are
 
 - Environment variables of the build (queried by `ci_is_env()`, `ci_has_env()` or `ci_get_env()`)
 - R Version of the current build (`getRversion()`)
 - other features of the build (e.g. branch name via `ci_get_branch()`)
 
-Most common tasks are the testing on multiple R versions and the restriction of certain tasks that should only be executed once (e.g. the [deployment of a _pkgdown_ site](deployment.html#pkgdown-deployment)).
+Most common tasks are the testing on multiple R versions and the restriction of certain tasks that should only be executed once (e.g. the [deployment of a {pkgdown} site](deployment.html#pkgdown-deployment)).
 
 ### Conditional execution: Use cases
 
@@ -254,5 +254,5 @@ knitr::include_graphics("img/circleci-debug.png")
 
 # What's not covered yet?
 
-- `SystemRequirements`: _tic_ is not yet capable of automatically determining system requirements specified in DESCRIPTION files of an R package. 
+- `SystemRequirements`: {tic} is not yet capable of automatically determining system requirements specified in DESCRIPTION files of an R package. 
 Our plan is to automatically provide suggestions like "Please add `addons: apt: <library>` to `.travis.yml`" in case system requirements are missing.
