+++ R/macro-pkgdown.R
@@ -62,6 +62,8 @@ do_pkgdown <- function(...,
     # account for old default "id_rsa"
     if (ci_has_env("id_rsa")) {
       name <- "id_rsa"
+    } else {
+      name <- travis_private_key_name
     }
     deploy <- ci_can_push(name = name)
 
