+++ R/appveyor.R
@@ -42,6 +42,8 @@ AppVeyorCI <- R6Class( # nolint
                          {.fun travis::use_travis_deploy} may help resolving
                          issues.", wrap = TRUE)
         stopc("This build cannot deploy to Github.")
+      } else {
+        return(can_push)
       }
     },
     get_env = function(env) {
