### 2020-01-12 c6c6b75
`step_install_ssh_keys()`: add support for old `id_rsa` default
cc @krlmlr: this should fix the tibble ci builds
