+++ R/steps-ssh.R
@@ -144,6 +144,10 @@ InstallSSHKeys <- R6Class(
 #'
 #' dsl_get()
 step_install_ssh_keys <- function(name = "TRAVIS_DEPLOY_KEY") {
+  # support for old "id_rsa" default: prefer this key if it exists
+  if (ci_has_env("id_rsa")) {
+    name <- "id_rsa"
+  }
   InstallSSHKeys$new(name = name)
 }
 
