+++ R/steps-install.R
@@ -52,18 +52,13 @@ InstallDeps <- R6Class(
 #' @export
 #' @name step_install_pkg
 #' @examples
-#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("install") %>%
 #'   add_step(step_install_deps())
 #'
 #' dsl_get()
-#' }
 step_install_deps <- function(repos = repo_default(), type = NULL) {
-  if (interactive()) {
-    stop("step_* functions should only be used in tic.R and not interactively.")
-  }
   type <- update_type(type)
   InstallDeps$new(repos = repos, type = type)
 }
@@ -110,19 +105,14 @@ InstallCRAN <- R6Class(
 #' @export
 #' @rdname step_install_pkg
 #' @examples
-#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("install") %>%
 #'   add_step(step_install_cran("magick"))
 #'
 #' dsl_get()
-#' }
 step_install_cran <- function(package = NULL, ..., repos = repo_default(),
                               type = NULL) {
-  if (interactive()) {
-    stop("step_* functions should only be used in tic.R and not interactively.")
-  }
   type <- update_type(type)
   InstallCRAN$new(package = package, repos = repos, ..., type = type)
 }
@@ -169,18 +159,13 @@ InstallGitHub <- R6Class(
 #' @export
 #' @rdname step_install_pkg
 #' @examples
-#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("install") %>%
 #'   add_step(step_install_github("rstudio/gt"))
 #'
 #' dsl_get()
-#' }
 step_install_github <- function(repo = NULL, ..., type = NULL) {
-  if (interactive()) {
-    stop("step_* functions should only be used in tic.R and not interactively.")
-  }
   type <- update_type(type)
   InstallGitHub$new(repo = repo, ..., type = type)
 }
