+++ R/dsl.R
@@ -24,11 +24,9 @@ NULL
 #' @rdname dsl
 #' @export
 #' @examples
-#' \dontrun{
 #' dsl_init()
 #'
 #' get_stage("script")
-#' }
 get_stage <- function(name) {
   # Initialize if necessary
   dsl_get()
@@ -48,12 +46,11 @@ get_stage <- function(name) {
 #' @rdname dsl
 #' @export
 #' @examples
-#' \dontrun{
+#'
 #' get_stage("script") %>%
 #'   add_step(step_hello_world())
 #'
 #' get_stage("script")
-#' }
 add_step <- function(stage, step) {
   step_quo <- enquo(step)
 
@@ -81,12 +78,11 @@ add_step <- function(stage, step) {
 #' @inheritParams step_run_code
 #' @rdname dsl
 #' @examples
-#' \dontrun{
+#'
 #' get_stage("script") %>%
 #'   add_code_step(print("Hi!"))
 #'
 #' get_stage("script")
-#' }
 add_code_step <- function(stage, call = NULL, prepare_call = NULL) {
   call_expr <- enexpr(call)
   prepare_call_expr <- enexpr(prepare_call)
