+++ man/dsl.Rd
@@ -50,21 +50,17 @@ and the links therein for available steps.
 \code{add_code_step()} is a shortcut for \code{add_step(step_run_code(...))}.
 }
 \examples{
-\dontrun{
 dsl_init()
 
 get_stage("script")
-}
-\dontrun{
+
 get_stage("script") \%>\%
   add_step(step_hello_world())
 
 get_stage("script")
-}
-\dontrun{
+
 get_stage("script") \%>\%
   add_code_step(print("Hi!"))
 
 get_stage("script")
-}
 }
