+++ man/do_package_checks.Rd
@@ -81,13 +81,11 @@ stages:
 }
 }
 \examples{
-\dontrun{
 dsl_init()
 
 do_package_checks()
 
 dsl_get()
-}
 }
 \seealso{
 Other macros: 
