+++ man/step_build_bookdown.Rd
@@ -46,12 +46,10 @@ supposed to be UTF-8.}
 Build a bookdown book using \code{\link[bookdown:render_book]{bookdown::render_book()}}.
 }
 \examples{
-\dontrun{
 dsl_init()
 
 get_stage("script") \%>\%
   add_step(step_build_bookdown("."))
 
 dsl_get()
-}
 }
