+++ tests/testthat/helper-git.R
@@ -1,3 +1,0 @@
-if (!dir.exists("tic")) {
-  system("git clone https://github.com/ropenscilabs/tic.git")
-}
