### 2020-01-12 da5988c
Show link to help page when creating a clean stage config
