+++ R/print.R
@@ -1,7 +1,7 @@
 #' @import cli
 #' @export
 print.TicStages <- function(x, ...) {
-  cli_alert_info("tic configuration")
+  cli_h1("tic configuration")
 
   if (all(vlapply(x, stage_is_empty))) {
     cat_bullet(
