### 2020-01-12 2648c5c
Check DESCRIPTION file for steps that need it
