+++ R/steps-pkgdown.R
@@ -20,6 +20,9 @@ BuildPkgdown <- R6Class(
     },
 
     prepare = function() {
+      if (!file.exists("DESCRIPTION")) {
+        stopc("No DESCRIPTION file found. The step_build_pkgdown step and the do_pkgdown macro are only available for packages.")
+      }
       verify_install(c("pkgdown", "remotes"))
       super$prepare()
     }
