+++ R/steps-install.R
@@ -10,7 +10,7 @@ InstallDeps <- R6Class(
 
     prepare = function() {
       if (!file.exists("DESCRIPTION")) {
-        stopc("No DESCRIPTION file found. The step_build_pkgdown step and the do_pkgdown macro are only available for packages.")
+        stopc("No DESCRIPTION file found. The step_install_deps step and the do_package_checks macro are only available for packages.")
       }
       verify_install("remotes")
     },
