### 2020-01-12 73d7ad6
Only use id_rsa if new key is not available, CC @pat-s
