+++ R/macro-bookdown.R
@@ -59,10 +59,9 @@ do_bookdown <- function(...,
     #'
     #'   1. The repo can be pushed to (see [ci_can_push()]).
     # account for old default "id_rsa"
-    if (ci_has_env("id_rsa")) {
+    name <- travis_private_key_name
+    if (ci_has_env("id_rsa") && !ci_has_env(name)) {
       name <- "id_rsa"
-    } else {
-      name <- travis_private_key_name
     }
     cli_text("Using {name} env var as the private key name for SSH deployment.")
     deploy <- ci_can_push(name = name)
@@ -87,10 +86,9 @@ do_bookdown <- function(...,
     #' 1. [step_setup_push_deploy()] in the `"before_deploy"` stage
     #'    (if `deploy` is set),
     #'
-    if (ci_has_env("id_rsa")) {
+    name <- travis_private_key_name
+    if (ci_has_env("id_rsa") && !ci_has_env(name)) {
       name <- "id_rsa"
-    } else {
-      name <- travis_private_key_name
     }
     get_stage("before_deploy") %>%
       add_step(step_setup_ssh(name = name)) %>%
