### 2020-01-12 0606e93
Umlaut, closes #207
