### 2020-01-12 5990a21
- `error_on = "note"` also fails on warnings.
CC @pat-s
