### 2020-01-13 ropenscilabs/tic#200 @krlmlr
Not sure we should require a `DESCRIPTION` for all projects.
