### 2020-01-13 #305 @krlmlr
`dsl_init()` doesn't require a `tic.R` file. In https://github.com/ropenscilabs/tic/pull/213 we have adapted the output after `dsl_init()` and macro functions such as `do_bookdown()` to include a link to a help page.

``` r
library(tic)

dsl_init()
#> ✓ Creating a clean tic stage configuration
#> ℹ See `?tic::dsl_get` for details
do_bookdown()
#> Using TRAVIS_DEPLOY_KEY env var as the private key name for SSH deployment.
#>
#> ── tic configuration ────────────────────────────────────────────────────────────────────────────────────
#> ── install ───────────────────────────────────────────────────────────────────────────────────── stage ──
#> ▶ step_install_deps(repos = repo_default())
#> ── deploy ────────────────────────────────────────────────────────────────────────────────────── stage ──
#> ▶ step_build_bookdown()
```
