### 2020-01-13 ropenscilabs/tic#200 @pat-s
How should `step_install_deps()` work otherwise?
