### 2020-01-13 ropenscilabs/tic#200 @krlmlr
`step_install_deps()` raises an error if `DESCRIPTION` is missing, but the other steps don't.
