### 2020-01-13 ropenscilabs/tic#200 @pat-s
Hm ok, I was just thinking that this step is basically within every pipeline and therefore we could quite early if DESC is missing (instead of waiting until the CI build arrives there and errors).
But there are just too many ways that one can get started with {tic} - we cannot safeguard all of them.

One could also argue why one would use {tic} without a proper pkg/project infrastructure.

I am with you that this is a very debatable change that also comes with some hurdles regarding the checking of this (which is imo also out of {tic}'s scope).
