### 2020-01-15 #305 @annakrystalli
Hello @pat-s and @krlmlr,

Thanks for all your work! I feel all my comments have been addressed apart from the following:

> what it is used for and how to set up R_TRAVIS token including an example of setting this up eg using usthis::edit_r_environ() to add env var to .Renviron

> I don't think that this belong into the {tic} package but rather into the {travis} package. The {tic} package has already so many vignettes and we need to take care not to bloat up the package too much. We can't write a vignette about the specialties of all existing and upcoming backends. There is already a section about how to set R_TRAVIS at https://docs.ropensci.org/travis/articles/travis.html#setting-the-default-endpoint.

Because this environmental variable needs to be set for `tic` to work,  it should also be mentioned in `tic` documentation. It's ok to just mention that it is required in the `tic` help file and getting started vignette and link to the `travis` vignette for more details if you prefer. But the `tic` user needs to know about it upfront. And remember, for documentation the opposite principle to code applies of [do repeat yourself](http://r-pkgs.had.co.nz/man.html#dry2) 😉

***

The rest are just some responses to some of your comments @pat-s and general suggestions but no action is necessary to pass review.

> I am not sure about such a vignette. Most people prefer searching error messages and not many people will read a vignette upfront before trying out the functions. Also I would not know right now what to put there.

As a user that tried to set up `tic` for docs in an organisation repo, I will have to disagree. Googling error messages was not fun and it took me ages to figure out how to set up authorisations and apps properly. The difficulties arise from how GitHub works (and their poor documentation IMHO) in that there are setting that need to be made on the org level AND on the user level, and the existence of the two endpoints and therefore travis apps on older users accounts further complicates the issue. So while it is not `tic`'s deficiency in any way and I will not hold up the review for this, my suggestion still stands to provide such information for users, ie at the very least, how to set up `tic` for an org repo.

> There is a small example in https://docs.ropensci.org/tic/articles/tic.html already. I am not sure what you mean by "not breaking down the questions"? We hope that ?tic::use_tic() helps. I am not sure to what extent a c/p tic.R in the README would add added value compared to all the contents in the already existing vignettes. The vignettes break down how {tic} works, what steps/stages are and that everything comes together in the tic.R file. Is there something you miss in all of these? Or could you maybe clarify more what the advantage of this action would be so I can follow?

Don't worry about this. I am likely being pedantic in that I felt it would make more sense to be explicit about the code being in a `tic.R` file rather than walking through tic code, just to hammer through that this is code that lives in this file and is not executed in the console. But just ignore this suggestion.

> I think users interested in a package simplifying CI usage are familiar enough with git/Github to know that their local repo needs to be linked. The error message is even quite descriptive imo. I don't think that checking this is within the responsibility of this package.

That's fine. The error message is ok. In general though I recommend you do not assume too much knowledge from your users, especially in technical terminology, (eg remotes) and try and help them debug as much as possible (which is why I made the suggestion). It will make your packages much nicer to work with for beginners.

For @krlmlr comment:
> In ropenscilabs/tic#213 we have adapted the output after dsl_init() and macro functions such as do_bookdown() to include a link to a help page
That's great.
> I wonder if the link to the help page should be printed every time a configuration is shown, or just once (as in the example above).
I don't have a strong opinion about this so will defer to what you think is best.

***

All in all we are pretty much there. If you could just address the one comment that still requires action we'll be ready for approval! 😃
