+++ R/local.R
@@ -4,15 +4,19 @@ LocalCI <- R6Class(
 
   public = list(
     get_branch = function() {
+      # Suppress warnings that occur if not in a Git repo
       suppressWarnings(system2("git", "rev-parse --abbrev-ref HEAD", stdout = TRUE))
     },
     get_tag = function() {
+      # Suppress warnings that occur if not in a Git repo
       suppressWarnings(system2("git", "describe", stdout = TRUE))
     },
     is_tag = function() {
+      # Suppress warnings that occur if not in a Git repo
       suppressWarnings(length(system2("git", c("tag", "--points-at", "HEAD"), stdout = TRUE)) > 0)
     },
     get_slug = function() {
+      # Suppress error that occurs if not in a Git repo
       tryCatch(
         {
           remote <- gh::gh_tree_remote()
@@ -28,6 +32,7 @@ LocalCI <- R6Class(
       NULL
     },
     get_commit = function() {
+      # Suppress error that occurs if not in a Git repo
       tryCatch(git2r::revparse_single(revision = "HEAD")$sha, error = "")
     },
     can_push = function(name = "TRAVIS_DEPLOY_KEY") {
