+++ R/steps-pkgdown.R
@@ -21,7 +21,8 @@ BuildPkgdown <- R6Class(
 
     prepare = function() {
       if (!file.exists("DESCRIPTION")) {
-        stopc("No DESCRIPTION file found. The step_build_pkgdown step and the do_pkgdown macro are only available for packages.")
+        cli_alert_danger("The {.code step_build_pkgdown()} step and the {.code do_pkgdown()} macro are only available for packages.")
+        stopc("No DESCRIPTION file found.")
       }
       verify_install(c("pkgdown", "remotes"))
       super$prepare()
