+++ R/use_tic.R
@@ -3,6 +3,10 @@
 #' @title Initialize CI testing using tic
 #' @description Prepares a repo for building and deploying supported by
 #' \pkg{tic}.
+#' Depending on your choice of CI system, this function requires
+#' properly set up authentication with Travis CI and/or Circle CI.
+#' See `vignette("travis", package = "travis")` and
+#' `vignette("circle", package = "circle")` for details.
 #'
 #' @importFrom utils menu
 #' @details
