### 2020-01-15 #305 @krlmlr
Thanks. I've added links to the relevant documentation for setup of environment variables to the `?use_tic` help page in https://github.com/ropenscilabs/tic/pull/213/commits/1bc46995dd26de5459e29e1650ff582a0e095108. I agree that this is out of scope for the package but we should provide helpful links. This form will be picked up by pkgdown's autolinker -- users of the pkgdown website will see a direct link to the relevant vignettes.
