+++ DESCRIPTION
@@ -34,8 +34,8 @@ Description: Provides a way to describe common build and
     recipe is described independent of the continuous integration tool
     used for processing the workflow (e.g.  'Travis CI' or 'AppVeyor').
 License: GPL (>= 2)
-URL: https://github.com/ropenscilabs/tic
-BugReports: https://github.com/ropenscilabs/tic/issues
+URL: https://github.com/ropensci/tic
+BugReports: https://github.com/ropensci/tic/issues
 Depends: 
     R (>= 3.2.0)
 Imports: 
