+++ DESCRIPTION
@@ -33,6 +33,7 @@ Description: Provides a way to describe common build and
     blogdown, pkgdown), or data processing (e.g. research compendia).  The
     recipe is described independent of the continuous integration tool
     used for processing the workflow (e.g.  'Travis CI' or 'AppVeyor').
+ This package has been peer-reviewed by rOpenSci (v. 0.3.0.9004).
 License: GPL (>= 2)
 URL: https://github.com/ropensci/tic
 BugReports: https://github.com/ropensci/tic/issues
