+++ DESCRIPTION
@@ -5,7 +5,7 @@ Title: Tasks Integrating Continuously: CI-Agnostic Workflow
 Version: 0.3.0.9004
 Authors@R: 
     c(person(given = "Kirill",
-             family = "M\u00fcller",
+             family = "Müller",
              role = c("aut", "cre"),
              email = "krlmlr+r@mailbox.org",
              comment = c(ORCID = "0000-0002-1416-3412")),
@@ -33,7 +33,7 @@ Description: Provides a way to describe common build and
     blogdown, pkgdown), or data processing (e.g. research compendia).  The
     recipe is described independent of the continuous integration tool
     used for processing the workflow (e.g.  'Travis CI' or 'AppVeyor').
- This package has been peer-reviewed by rOpenSci (v. 0.3.0.9004).
+    This package has been peer-reviewed by rOpenSci (v. 0.3.0.9004).
 License: GPL (>= 2)
 URL: https://github.com/ropensci/tic
 BugReports: https://github.com/ropensci/tic/issues
