### 2020-01-16 fb16d42
first do_drat() try
