+++ man/do_bookdown.Rd
@@ -104,6 +104,7 @@ dsl_get()
 }
 \seealso{
 Other macros: 
+\code{\link{do_drat}()},
 \code{\link{do_package_checks}()},
 \code{\link{do_pkgdown}()}
 }
