### 2020-01-16 43952e8
require full repo slug
