### 2020-01-16 ff245db
fix ci_can_push()
