### 2020-01-16 3006dc5
clean add_to_drat()
