### 2020-01-16 #305 @annakrystalli
Great, thanks @pat-s !

It looks like you and @krlmlr still have full admin rights so I think we're all done here and I can close the issue.

I also think it would be great to announce the package to the rOpenSci community so do consider writing a blogpost/technical note if you have time/the inclination ☺️. Just let @stefaniebutland know if you are interested.
